### 2025-06-20 #8437 @editorialbot
[PRE REVIEW]: SpatialProteomicsNet: A unified interface for spatial proteomics data access for computer vision and machine learning
**Submitting author:** @adrianomartinelli (<a href="http://orcid.org/0000-0002-9288-5103">Adriano Martinelli</a>)
**Repository:** https://github.com/AI4SCR/ai4bmr-datasets
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
