### 2025-06-21 #8437 @adrianomartinelli
Dear editors,

not sure why the pdf compilation errors. I used the workflow configuration provided by JOSS in my repo, which compiles.

```
  steps:
    - name: Checkout
      uses: actions/checkout@v4
    - name: Build draft PDF
      uses: openjournals/openjournals-draft-action@master
      with:
        journal: joss
        # This should be the path to the paper within your repo.
        paper-path: paper.md
```

However, the failed run used `xuanxu/paper-action@main`.
