### 2025-06-30 0dcd7d7
improve tiff preparation and draft flag post processing
