+++ tests/datasets/test_BEAT.py
@@ -1,5 +1,7 @@
-def test_Keren2018():
+def test_BEAT():
     from ai4bmr_datasets import BEAT
     ds = self = BEAT()
-    ds.prepare_data()
-    ds.setup()
+    ds.prepare_tools()
+    ds.prepare_clinical()
+    ds.prepare_wsi(force=True)
+    # ds.setup()
