+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -13,7 +13,6 @@ class BEAT:
         self.processed_dir = self.base_dir / '02_processed'
         self.datasets_dir = self.processed_dir / 'datasets'
         self.tools_dir = self.base_dir / '03_tools'
-        self.log_dir = self.base_dir / '04_logs'
 
     def prepare_tools(self):
         bfconvert = self.tools_dir / "bftools" / "bfconvert"
@@ -124,7 +123,6 @@ class BEAT:
         import subprocess
 
         self.prepare_tools()
-        self.log_dir.mkdir(parents=True, exist_ok=True)
 
         dataset_dir = self.datasets_dir / 'beat'
         dataset_dir.mkdir(parents=True, exist_ok=True)
@@ -142,17 +140,18 @@ class BEAT:
             if save_path.exists() and not force:
                 logger.info(f"File {save_path} already exists. Skipping conversion.")
 
-            job_name = f"wsi2tiff-{save_name}"
+            job_name = f"proj=beat-task=wsi2tiff-{save_name}"
             logger.info(f"Converting {wsi_path} to {save_path} using Bio-Formats (job_name={job_name})")
+            num_cpus = 16
 
             sbatch_command = f"""
             sbatch \\
             --job-name={job_name} \\
-            --output=/users/amarti51/logs/{job_name}-%A.log \\
-            --error=/users/amarti51/logs/{job_name}-%A.err \\
-            --time=12:00:00 \\
+            --output=/work/FAC/FBM/DBC/mrapsoma/prometex/logs/adrianom/{job_name}-%j.log \\
+            --error=/work/FAC/FBM/DBC/mrapsoma/prometex/logs/adrianom/{job_name}-%j.err \\
+            --time=24:00:00 \\
             --mem=128G \\
-            --cpus-per-task=12 \\
+            --cpus-per-task={num_cpus} \\
             <<'EOF'
             #!/bin/bash
             
@@ -167,9 +166,13 @@ class BEAT:
             mkdir -p "$(dirname "{save_path}")"
 
             # Convert NDPI/other WSI to intermediate TIFF
-            /usr/bin/time -v "$BFCONVERT" -nogroup -bigtiff -compression LZW -tilex 512 -tiley 512 -series 0 -pyramid-resolutions 4 -pyramid-scale 2 "$INPUT" "$INTERMEDIATE"
+            # /usr/bin/time -v "$BFCONVERT" -nogroup -bigtiff -compression LZW -tilex 512 -tiley 512 -series 0 -pyramid-resolutions 4 -pyramid-scale 2 "$INPUT" "$INTERMEDIATE"
+            echo "* Exporting full-res TIFF..."
+            /usr/bin/time -v "$BFCONVERT" -nogroup -bigtiff -series 0 "$INPUT" "$INTERMEDIATE"
 
             # Convert intermediate TIFF to OpenSlide-compatible pyramid TIFF
+            echo "* Building OpenSlide-compatible pyramid TIFF..."
+            export VIPS_CONCURRENCY={num_cpus - 1}
             /usr/bin/time -v vips tiffsave "$INTERMEDIATE" "$OUTPUT" --pyramid --tile --tile-width 512 --tile-height 512 --bigtiff --compression lzw
             
             rm "$INTERMEDIATE"
@@ -179,6 +182,17 @@ class BEAT:
             sbatch_command = textwrap.dedent(sbatch_command).strip()
 
             subprocess.run(sbatch_command, shell=True, check=True)
+
+    def post_process_tiff_flags(self):
+        # # set XResolution (tag 282) to 22611
+        # tiffset -s 282 22611 "$OUTPUT"
+        #
[…]
