### 2025-07-01 2d7241e
create thumbnails and adjust tiff flags
