### 2025-07-01 080ed10
create patch embeddigns
