### 2025-07-01 3bb359a
configure batch size and num workers
