### 2025-07-02 ddaf4dc
debug filter coords
