### 2025-07-15 #8437 @adrianomartinelli
Dear @crvernon ,

I hope you’re well. I wanted to follow up to see if there have been any updates regarding the discussion with the editorial board about our submission.

We believe the package would be a valuable resource for researchers developing models on spatial proteomics datasets, and we’d be happy to see it published in JOSS.

Best regards,
Adriano
