### 2025-07-16 d5cc5af
ChEMBL
