+++ src/ai4bmr_datasets/datasets/__init__.py
@@ -3,6 +3,7 @@ from .Keren2018 import Keren2018
 from .Cords2024 import Cords2024
 from .Danenberg import Danenberg2022
 from .Jackson2020 import Jackson2020
+from .ChEMBL import ChEMBL
 
 from .DummyPoints import DummyPoints
 from .DummyImages import DummyImages
