+++ src/ai4bmr_datasets/utils/download.py
@@ -1,5 +1,8 @@
 from pathlib import Path
 
+from loguru import logger
+
+
 def unzip_recursive(zip_path: Path, extract_dir: Path = None):
     import zipfile
     from loguru import logger
@@ -25,3 +28,31 @@ def unzip_recursive(zip_path: Path, extract_dir: Path = None):
             extracted_path = extract_dir / name
             if extracted_path.suffix == ".zip":
                 unzip_recursive(extracted_path)
+
+
+def download_file_map(file_map: dict[str, Path], force: bool = False):
+    from tqdm import tqdm
+    import requests
+
+    for url, target_path in file_map.items():
+        if not target_path.exists() or force:
+            logger.info(f"Downloading {url} → {target_path}")
+            headers = {"User-Agent": "Mozilla/5.0"}
+            with requests.get(url, headers=headers, stream=True) as r:
+                r.raise_for_status()
+                total_size = int(r.headers.get('Content-Length', 0))
+                chunk_size = 8192
+
+                with open(target_path, 'wb') as f, tqdm(
+                        desc=target_path.name,
+                        total=total_size,
+                        unit='B',
+                        unit_scale=True,
+                        unit_divisor=1024,
+                ) as pbar:
+                    for chunk in r.iter_content(chunk_size=chunk_size):
+                        if chunk:  # filter out keep-alive chunks
+                            f.write(chunk)
+                            pbar.update(len(chunk))
+        else:
+            logger.info(f"Skipping download of {target_path.name}, already exists.")
