+++ src/ai4bmr_datasets/datasets/ChEMBL.py
@@ -18,8 +18,17 @@ class ChEMBL:
     def prepare_data(self):
         self.download()
 
-    def setup(self):
+    def setup(self, query: str | None = None):
         import sqlite3
+
+        query = query or """SELECT
+    m.chembl_id, 
+    s.canonical_smiles
+FROM molecule_dictionary m
+JOIN compound_structures s ON m.molregno = s.molregno
+WHERE s.canonical_smiles IS NOT NULL
+"""
+
         if self.db is None:
             db_path = self.sqlite_file
             if not db_path.exists():
@@ -27,17 +36,7 @@ class ChEMBL:
 
             logger.info(f"Loading compound_structures from {db_path}")
             conn = sqlite3.connect(db_path)
-            self.db = pd.read_sql_query(
-                """
-                SELECT 
-                    m.chembl_id, 
-                    s.canonical_smiles
-                FROM molecule_dictionary m
-                JOIN compound_structures s ON m.molregno = s.molregno
-                WHERE s.canonical_smiles IS NOT NULL
-                """,
-                conn
-            )
+            self.db = pd.read_sql_query(query, conn)
             conn.close()
 
     def download(self, force: bool = False):
@@ -60,5 +59,4 @@ class ChEMBL:
         return len(self.db)
 
     def __getitem__(self, idx: int) -> dict:
-        row = self.db.iloc[idx]
-        return dict(chembl_id = row["chembl_id"], smiles=row["canonical_smiles"])
+        return self.db.iloc[idx].to_dict()
