### 2025-07-30 4b073ce
count nan
