### 2025-08-06 #8437 @ujjwalkarn
@sumn2u @ggsharma @surajpaib @Timozen would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/review_criteria.htm](https://joss.readthedocs.io/en/latest/review_criteria.html)
