### 2025-08-08 b51dfa3
Refactor: Move dataset configuration to __init__
This refactoring moves the dataset configuration from the `setup` method to the `__init__` method of the `BaseIMCDataset` class and its subclasses.

This change makes the dataset classes more compatible with downstream deep learning frameworks like PyTorch Lightning, which expect to be able to instantiate dataset objects without immediately loading all the data.
