+++ tests/datasets/test_Cords2024.py
@@ -14,11 +14,11 @@ def test_Cords2024():
     from ai4bmr_datasets import Cords2024 as Dataset
 
     image_version = mask_version = metadata_version = feature_version = 'published'
-    ds = self = Dataset()
+    ds = self = Dataset(image_version=image_version, mask_version=mask_version,
+                         metadata_version=metadata_version, load_metadata=True,
+                         feature_version=feature_version, load_intensity=True, load_spatial=True)
     # ds.create_annotated()
-    ds.setup(image_version=image_version, mask_version=mask_version,
-             metadata_version=metadata_version, load_metadata=True,
-             feature_version=feature_version, load_intensity=True, load_spatial=True)
+    ds.setup()
 
     assert isinstance(ds.images, dict)
     assert isinstance(ds.masks, dict)
