+++ tests/datasets/test_Jackson2020.py
@@ -4,10 +4,11 @@ def test_Jackson2020():
     datasets_dir = Path('/work/FAC/FBM/DBC/mrapsoma/prometex/data/datasets/')
 
     image_version = mask_version = metadata_version = feature_version = 'published'
-    ds = self = Dataset(base_dir=datasets_dir / Dataset.name)
-    ds.setup(image_version=image_version, mask_version=mask_version,
-             metadata_version=metadata_version, load_metadata=True,
-             feature_version=feature_version, load_intensity=True, load_spatial=True)
+    ds = self = Dataset(base_dir=datasets_dir / Dataset.name,
+                         image_version=image_version, mask_version=mask_version,
+                         metadata_version=metadata_version, load_metadata=True,
+                         feature_version=feature_version, load_intensity=True, load_spatial=True)
+    ds.setup()
 
     assert isinstance(ds.images, dict)
     assert isinstance(ds.masks, dict)
