### 2025-08-08 c87cec9
Docs: Add docstrings
