### 2025-08-08 b4caa67
Docs: Add docstrings
