+++ src/ai4bmr_datasets/datasets/BaseIMCDataset.py
@@ -11,11 +11,19 @@ class BaseIMCDataset:
     doi: str = None
 
     """
-    Base class for managing Imaging Mass Cytometry (IMC) datasets,
-    including loading images, masks, features (intensity, spatial), and metadata.
+    Base class for managing Imaging Mass Cytometry (IMC) datasets.
+
+    This class provides a standardized interface for loading and processing IMC data,
+    including images, masks, features (intensity, spatial), and clinical metadata.
+    It handles directory resolution, data versioning, and alignment of different data modalities.
+
+    Attributes:
+        name (str): The name of the dataset.
+        id (str): A unique identifier for the dataset.
+        doi (str): The Digital Object Identifier (DOI) for the dataset's publication.
     """
 
-    def __init__(self, 
+    def __init__(self,
                  base_dir: Path | None = None,
                  image_version: str | None = None,
                  mask_version: str | None = None,
@@ -28,10 +36,20 @@ class BaseIMCDataset:
                  join: str = "outer",
                  ):
         """
-        Initialize the dataset with a base directory.
+        Initializes the BaseIMCDataset with specified configurations.
 
         Args:
-            base_dir (Path, None): Base directory for the dataset that contains 01_raw and 02_processed folders.
+            base_dir (Path | None): Base directory for the dataset, containing '01_raw' and '02_processed' folders.
+                                    If None, it defaults to an environment variable or a cache directory.
+            image_version (str | None): Identifier for the image version to load.
+            mask_version (str | None): Identifier for the mask version to load.
+            feature_version (str | None): Identifier for the feature version to load.
+            metadata_version (str | None): Identifier for the metadata version to load.
+            load_intensity (bool): If True, intensity features will be loaded.
+            load_spatial (bool): If True, spatial features will be loaded.
+            load_metadata (bool): If True, observation-level metadata will be loaded.
+            align (bool): If True, aligns and intersects all available sample IDs across modalities.
+            join (str): Specifies the type of join to perform during alignment (e.g., 'outer').
         """
 
         self.base_dir = self.resolve_base_dir(base_dir=base_dir)
@@ -54,7 +72,19 @@ class BaseIMCDataset:
         self.intensity = None
         self.spatial = None
 
-    def resolve_base_dir(self, base_dir: Path | None):
+    def resolve_base_dir(self, base_dir: Path | None) -> Path:
+        """
+        Resolves the base directory for the dataset.
+
+        If `base_dir` is None, it attempts to use the 'AI4BMR_DATASETS_DIR' environment variable.
+        If the environment variable is not set, it defaults to a cache directory within the user's home directory.
+
+        Args:
+            base_dir (Path | None): The initial base directory provided.
+
+        Returns:
+            Path: The resolved absolute path to the base directory.
+        """
         import os
         if base_dir is None:
             base_dir = os.environ.get("AI4BMR_DATASETS_DIR", None)
@@ -64,10 +94,26 @@ class BaseIMCDataset:
                 base_dir = Path(base_dir) / self.name
         return base_dir
 
-    def __len__(self):
+    def __len__(self) -> int:
+        """
+        Returns the number of samples in the dataset.
+
+        Returns:
+            int: The number of samples.
+        """
         return len(self.sample_ids)
 
-    def __getitem__(self, idx):
+    def __getitem__(self, idx: int) -> dict:
+        """
+        Retrieves a single sample from the dataset by index.
+
+        Args:
+            idx (int): The index of the sample to retrieve.
+
+        Returns:
+            dict: A dictionary containing the sample's ID, image, mask, panel, intensity,
+                  spatial features, and metadata.
+        """
         sample_id = self.sample_ids[idx]
[…]

[…]
