### 2025-08-08 5ba70b7
Revert "Docs: Add docstrings"
This reverts commit c87cec9a6f50ad6466e011f006e2af383407e5de.
