+++ src/ai4bmr_datasets/datasets/Jackson2020.py
@@ -38,10 +38,6 @@ class Jackson2020(BaseIMCDataset):
                          join=join)
 
     def prepare_data(self):
-        """
-        Prepares the Jackson2020 dataset by downloading, creating clinical metadata,
-        panel, images, masks, and features, and finally creating annotated data.
-        """
         self.download()
 
         self.create_clinical_metadata()
@@ -54,12 +50,6 @@ class Jackson2020(BaseIMCDataset):
         self.create_annotated()
 
     def download(self, force: bool = False):
-        """
-        Downloads the raw data files for the Jackson2020 dataset from Zenodo.
-
-        Args:
-            force (bool): If True, forces re-download even if files already exist.
-        """
         import requests
         import shutil
 
@@ -96,12 +86,6 @@ class Jackson2020(BaseIMCDataset):
         logger.info("✅ Download and extraction completed.")
 
     def create_images(self):
-        """
-        Creates and saves processed image files for the Jackson2020 dataset.
-
-        This method reads raw OME-TIFF images, filters them based on the panel,
-        and saves the processed images as TIFF files.
-        """
         panel = pd.read_parquet(self.get_panel_path('published'))
         clinical = pd.read_parquet(self.clinical_metadata_path)
 
@@ -130,12 +114,6 @@ class Jackson2020(BaseIMCDataset):
                 logger.warning(f"Image file {img_path} does not exist. Skipping.")
 
     def create_masks(self):
-        """
-        Creates and saves processed mask files for the Jackson2020 dataset.
-
-        This method reads raw mask files, processes them to match image dimensions,
-        and saves the processed masks as TIFF files.
-        """
         clinical = pd.read_parquet(self.clinical_metadata_path)
 
         raw_masks_dir = self.raw_dir / 'ome_and_single_cell_masks/OMEnMasks/Basel_Zuri_masks/Basel_Zuri_masks'
@@ -163,12 +141,6 @@ class Jackson2020(BaseIMCDataset):
                 logger.warning(f"Mask file {mask_path} does not exist. Skipping.")
 
     def create_metadata(self):
-        """
-        Creates and saves processed metadata (cell cluster labels) for the Jackson2020 dataset.
-
-        This method reads raw metacluster labels from Basel and Zurich cohorts,
-        merges them, and saves the processed metadata as Parquet files, grouped by sample ID.
-        """
         bs_meta_clusters = pd.read_csv(
             self.raw_dir / 'single_cell_cluster_labels/Cluster_labels/Basel_metaclusters.csv')
         zh_meta_clusters = pd.read_csv(
@@ -255,16 +227,7 @@ class Jackson2020(BaseIMCDataset):
             save_path = save_dir / f"{grp_name}.parquet"
             grp_dat.to_parquet(save_path, engine='fastparquet')
 
-    def add_sample_id_and_object_id(self, data: pd.DataFrame) -> pd.DataFrame:
-        """
-        Adds 'sample_id' and 'object_id' columns to the input DataFrame based on 'core' and 'CellId'.
-
-        Args:
-            data (pd.DataFrame): The input DataFrame containing 'core' and optionally 'CellId'.
-
-        Returns:
-            pd.DataFrame: The DataFrame with 'sample_id' and 'object_id' set as part of the index.
-        """
+    def add_sample_id_and_object_id(self, data):
         sample_ids = data.core.str.replace('ZTMA208_slide_', '')
         sample_ids = sample_ids.str.replace('BaselTMA_SP', '')
 
@@ -279,13 +242,6 @@ class Jackson2020(BaseIMCDataset):
         return data
 
     def create_clinical_metadata(self):
-        """
-        Creates and saves the clinical metadata for the Jackson2020 dataset.
-
[…]
