+++ src/ai4bmr_datasets/datasets/Cords2024.py
@@ -16,19 +16,13 @@ class Cords2024(BaseIMCDataset):
 
     @property
     def id(self):
-        """
-        Returns the unique identifier for the Cords2024 dataset.
-        """
         return "Cords2024"
 
     @property
     def doi(self):
-        """
-        Returns the DOI for the Cords2024 dataset publication.
-        """
         return "10.1016/j.ccell.2023.12.021"
 
-    def __init__(self,
+    def __init__(self, 
                  base_dir: Path | None = None,
                  image_version: str | None = None,
                  mask_version: str | None = None,
@@ -40,21 +34,6 @@ class Cords2024(BaseIMCDataset):
                  align: bool = False,
                  join: str = "outer",
                  ):
-        """
-        Initializes the Cords2024 dataset.
-
-        Args:
-            base_dir (Path | None): Base directory for the dataset.
-            image_version (str | None): Version of the images to load.
-            mask_version (str | None): Version of the masks to load.
-            feature_version (str | None): Version of the features to load.
-            metadata_version (str | None): Version of the metadata to load.
-            load_intensity (bool): Whether to load intensity features.
-            load_spatial (bool): Whether to load spatial features.
-            load_metadata (bool): Whether to load metadata.
-            align (bool): Whether to align sample IDs across modalities.
-            join (str): Type of join for alignment (e.g., 'outer').
-        """
 
         super().__init__(base_dir=base_dir,
                          image_version=image_version,
@@ -72,10 +51,6 @@ class Cords2024(BaseIMCDataset):
         self.raw_acquisitions_dir = self.raw_dir / "acquisitions"
 
     def prepare_data(self):
-        """
-        Prepares the Cords2024 dataset by downloading, processing, and creating
-        various data components (clinical metadata, panel, images, masks, metadata).
-        """
         self.download()
         self.process_rdata()
         self.process_acquisitions()
@@ -90,12 +65,6 @@ class Cords2024(BaseIMCDataset):
         self.create_annotated()
 
     def download(self, force: bool = False):
-        """
-        Downloads the raw data files for the Cords2024 dataset from Zenodo and Google Drive.
-
-        Args:
-            force (bool): If True, forces re-download even if files already exist.
-        """
         import requests
         import shutil
 
@@ -169,25 +138,23 @@ class Cords2024(BaseIMCDataset):
         logger.info("✅ Download and extraction completed.")
 
     def get_sample_id_from_path(self, path):
-        """
-        Extracts the sample ID from a given file path.
-
-        Args:
-            path (str): The file path.
-
-        Returns:
-            str: The extracted sample ID.
-        """
         path = str(path)
         search = re.search(r"TMA_(\d+)_(\w+)_s0_a(\d+)", (path))
         return '_'.join(search.groups())
 
     def process_acquisitions(self):
         """
-        Processes raw MCD acquisition files to extract and save image data and metadata.
+        Failed reading acquisitions for:
+        # have masks
+        - 88_A_10
+        - 88_A_9
+        - 88_B_110
+        - 86_B_1
+        - 176_B_47
+        - 175_B_90
 
-        This method reads .mcd files, extracts individual acquisitions, and saves them as
-        TIFF images and JSON metadata files in the raw acquisitions directory.
+        - 176_C_2
+        - 87_A_49
         """
 
         from readimc import MCDFile
@@ -239,13 +206,6 @@ class Cords2024(BaseIMCDataset):
                             continue
 
     def validate_masks(self):
-        """
-        Validates the consistency between mask objects and metadata objects.
-
-        This method loads published image, mask, and metadata versions, then checks if the unique
[…]

[…]
