+++ src/ai4bmr_datasets/datasets/Sorin2023.py
@@ -1,0 +1,657 @@
+import json
+import re
+from pathlib import Path
+from tqdm import tqdm
+import numpy as np
+import pandas as pd
+from loguru import logger
+from tifffile import imread, imwrite
+
+from ai4bmr_datasets.datasets.BaseIMCDataset import BaseIMCDataset
+from ai4bmr_datasets.utils.download import unzip_recursive
+
+
+class Sorin2023(BaseIMCDataset):
+    name = "Sorin2023"
+
+    @property
+    def id(self):
+        return "Sorin2023"
+
+    @property
+    def doi(self):
+        return "10.1038/s41586-022-05672-3"
+
+    def __init__(self, base_dir: Path | None = None):
+        super().__init__(base_dir)
+
+    def prepare_data(self):
+        self.download()
+        self.process_rdata()
+        self.process_acquisitions()
+
+        self.create_clinical_metadata()
+
+        self.create_panel()
+        self.create_images()
+        self.create_masks()
+        self.create_metadata()
+
+        self.create_annotated()
+
+    def download(self, force: bool = False):
+        import requests
+        import shutil
+
+        download_dir = self.raw_dir
+        download_dir.mkdir(parents=True, exist_ok=True)
+
+        file_map = {
+            "https://zenodo.org/records/7961844/files/2020115_LC_NSCLC_TMA_86_A.mcd.zip?download=1": download_dir
+                                                                                                     / "2020115_LC_NSCLC_TMA_86_A.mcd.zip",
+            "https://zenodo.org/records/7961844/files/2020117_LC_NSCLC_TMA_86_B.mcd.zip?download=1": download_dir
+                                                                                                     / "2020117_LC_NSCLC_TMA_86_B.mcd.zip",
+            "https://zenodo.org/records/7961844/files/2020120_LC_NSCLC_TMA_86_C.mcd.zip?download=1": download_dir
+                                                                                                     / "2020120_LC_NSCLC_TMA_86_C.mcd.zip",
+            "https://zenodo.org/records/7961844/files/20201219_LC_NSCLC_TMA_178_A.mcd.zip?download=1": download_dir
+                                                                                                       / "20201219_LC_NSCLC_TMA_178_A.mcd.zip",
+            "https://zenodo.org/records/7961844/files/20201222_LC_NSCLC_TMA_178_B_2.mcd.zip?download=1": download_dir
+                                                                                                         / "20201222_LC_NSCLC_TMA_178_B_2.mcd.zip",
+            "https://zenodo.org/records/7961844/files/20201224_LC_NSCLC_TMA_178_C.mcd.zip?download=1": download_dir
+                                                                                                       / "20201224_LC_NSCLC_TMA_178_C.mcd.zip",
+            "https://zenodo.org/records/7961844/files/20201226_LC_NSCLC_TMA_175_A.mcd.zip?download=1": download_dir
+                                                                                                       / "20201226_LC_NSCLC_TMA_175_A.mcd.zip",
+            "https://zenodo.org/records/7961844/files/20201228_LC_NSCLC_TMA_175_B.mcd.zip?download=1": download_dir
+                                                                                                       / "20201228_LC_NSCLC_TMA_175_B.mcd.zip",
+            "https://zenodo.org/records/7961844/files/20210101_LC_NSCLC_TMA_175_C.mcd.zip?download=1": download_dir
+                                                                                                       / "20210101_LC_NSCLC_TMA_175_C.mcd.zip",
+            "https://zenodo.org/records/7961844/files/20210104_LC_NSCLC_TMA_176_A.mcd.zip?download=1": download_dir
[…]
