### 2025-08-08 30043ac
use tempdir context manager for tests
