### 2025-08-08 860203f
Refactor: Harmonize download functions
