+++ src/ai4bmr_datasets/datasets/Danenberg.py
@@ -32,8 +32,8 @@ class Danenberg2022(BaseIMCDataset):
         self.create_annotated(mask_version='published_nucleus')
 
     def download(self, force: bool = False):
-        import requests
         import shutil
+        from ai4bmr_datasets.utils.download import download_file_map, unzip_recursive
 
         download_dir = self.raw_dir
         download_dir.mkdir(parents=True, exist_ok=True)
@@ -43,18 +43,7 @@ class Danenberg2022(BaseIMCDataset):
             'https://zenodo.org/records/15615709/files/correctedPublicMasks.zip?download=1': download_dir / 'corrected_public_masks.zip',
         }
 
-        # Download files
-        for url, target_path in file_map.items():
-            if not target_path.exists() or force:
-                logger.info(f"Downloading {url} → {target_path}")
-                headers = {"User-Agent": "Mozilla/5.0"}
-                with requests.get(url, headers=headers, stream=True) as r:
-                    r.raise_for_status()
-                    with open(target_path, 'wb') as f:
-                        for chunk in r.iter_content(chunk_size=8192):
-                            f.write(chunk)
-            else:
-                logger.info(f"Skipping download of {target_path.name}, already exists.")
+        download_file_map(file_map, force=force)
 
         # Extract zip files
         for target_path in file_map.values():
