### 2025-08-08 e02715b
CI: Configure GitHub Pages deployment and fix Cords2024.py syntax
