### 2025-08-08 9843d62
Revert: Cords2024.py changes from previous commit
