+++ src/ai4bmr_datasets/datasets/__init__.py
@@ -6,7 +6,3 @@ from .Jackson2020 import Jackson2020
 from .PCa import PCa
 
 from .BEAT import BEAT
-
-# from .DummyPoints import DummyPoints
-# from .DummyImages import DummyImages
-# from .DummyTabular import DummyTabular
