+++ .github/workflows/deploy-docs.yml
@@ -33,7 +33,9 @@ jobs:
 
     # Build the HTML documentation into docs/build/html using the Makefile.
     - name: Build documentation
-      run: make -C docs html
+      run: |
+        export PYTHONPATH=$PYTHONPATH:$(pwd)/src
+        sphinx-build -b html docs/source docs/build/html
 
     # Prepare GitHub Pages deployment (creates the gh-pages branch and sets up authentication).
     - uses: actions/configure-pages@v5
