### 2025-08-08 74e12ed
Docs: Add Installation section and restructure API docs
