+++ docs/source/installation.rst
@@ -1,0 +1,16 @@
+Installation
+============
+
+To install the `ai4bmr-datasets` package, you can use pip:
+
+.. code-block:: bash
+
+    pip install ai4bmr-datasets
+
+For development, you can clone the repository and install in editable mode:
+
+.. code-block:: bash
+
+    git clone https://github.com/AI4SCR/ai4bmr-datasets.git
+    cd ai4bmr-datasets
+    pip install -e .
