+++ docs/source/api.rst
@@ -1,0 +1,8 @@
+API Documentation
+=================
+
+.. automodule:: ai4bmr_datasets
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :no-index:
