+++ docs/source/datasets/Danenberg.rst
@@ -1,0 +1,23 @@
+Danenberg Dataset
+=================
+
+.. currentmodule:: ai4bmr_datasets.datasets.Danenberg
+
+.. autoclass:: Danenberg2022
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :exclude-members: name, id, doi
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
+
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
