+++ docs/source/datasets/Keren2018.rst
@@ -1,0 +1,23 @@
+Keren2018 Dataset
+=================
+
+.. currentmodule:: ai4bmr_datasets.datasets.Keren2018
+
+.. autoclass:: Keren2018
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :exclude-members: name, id, doi
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
+
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
