+++ docs/source/datasets/Jackson2020.rst
@@ -1,0 +1,23 @@
+Jackson2020 Dataset
+===================
+
+.. currentmodule:: ai4bmr_datasets.datasets.Jackson2020
+
+.. autoclass:: Jackson2020
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :exclude-members: name, id, doi
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
+
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
