+++ docs/source/datasets/BEAT.rst
@@ -1,0 +1,11 @@
+BEAT Dataset
+============
+
+.. currentmodule:: ai4bmr_datasets.datasets.BEAT
+
+.. autoclass:: BEAT
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :exclude-members: name
+
