+++ docs/source/datasets/Cords2024.rst
@@ -1,0 +1,23 @@
+Cords2024 Dataset
+=================
+
+.. currentmodule:: ai4bmr_datasets.datasets.Cords2024
+
+.. autoclass:: Cords2024
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :exclude-members: name, id, doi
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
+
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
