+++ docs/source/datasets/PCa.rst
@@ -1,0 +1,23 @@
+PCa Dataset
+===========
+
+.. currentmodule:: ai4bmr_datasets.datasets.PCa
+
+.. autoclass:: PCa
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :exclude-members: name, id, doi
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
+
+
+   .. autoproperty:: id
+      :no-index:
+
+   .. autoproperty:: doi
+      :no-index:
