### 2025-08-08 ce2a264
Docs: Add overview section and badges to README
