+++ docs/source/index.rst
@@ -15,6 +15,7 @@ documentation for details.
    :maxdepth: 2
    :caption: Contents:
 
+   overview
    installation
    api
 
