+++ docs/source/api.rst
@@ -1,8 +1,7 @@
 API Documentation
 =================
 
-.. automodule:: ai4bmr_datasets
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :no-index:
+.. toctree::
+   :maxdepth: 2
+
+   datasets
