+++ docs/build/html/_sources/index.rst.txt
@@ -15,8 +15,9 @@ documentation for details.
    :maxdepth: 2
    :caption: Contents:
 
-.. automodule:: ai4bmr_datasets
-   :members:
-   :undoc-members:
-   :show-inheritance:
+   overview
+   installation
+   api
+
+
 
