+++ src/ai4bmr_datasets/datasets/Cords2024.py
@@ -291,8 +291,7 @@ class Cords2024(BaseIMCDataset):
         panel["target"] = targets.str[0].map(tidy_name).values
         # add suffix to target for duplicates (iridium)
         panel['target'] = panel.target + '_' + panel.groupby('target').cumcount().astype(str)
-        panel.loc[:, 'target'] = panel.target.str.replace('_0
-, '', regex=True)
+        panel.loc[:, 'target'] = panel.target.str.replace('_0', '', regex=True)
 
         panel["keep"] = keep
 
