### 2025-08-23 ae6be51
refactor into more granular tasks
