+++ slurm/scripts/compute-beat-embeddings.py
@@ -1,3 +1,3 @@
 from ai4bmr_datasets import BEAT
 dm = BEAT()
-dm.create_patch_embeddings(model_name='uni_v1', num_workers=12, batch_size=64)
+dm.create_embeddings(model_name='uni_v1', num_workers=12, batch_size=64)
