+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -1,9 +1,30 @@
-import shutil
-from openslide import OpenSlide
+from pathlib import Path
+import json
+import subprocess
+import textwrap
+from dataclasses import asdict
 from pathlib import Path
 
+import geopandas as gpd
+import numpy as np
+import openslide
 import pandas as pd
+import torch
+from ai4bmr_core.utils.tidy import tidy_name
+from ai4bmr_learn.data_models.Coordinate import SlideCoordinate
+from ai4bmr_learn.datasets.items import Patches, get_patch
+from ai4bmr_learn.plotting.contours import visualize_contours
+from ai4bmr_learn.plotting.patches import visualize_coords
+from ai4bmr_learn.utils.images import filter_coords
+from ai4bmr_learn.utils.slides import get_coordinates_dict, get_mpp, get_mpp_and_resolution, get_seg_model, \
+    get_slide_patcher_params, get_thumbnail, segment_slide
 from loguru import logger
+from openslide import OpenSlide
+from skimage.io import imsave
+from torch.utils.data import DataLoader
+from torchvision.transforms import v2
+from tqdm import tqdm
+from trident.patch_encoder_models import encoder_factory, encoder_factory
 
 
 class BEAT:
@@ -22,8 +43,6 @@ class BEAT:
             logger.info(f"bfconvert found at {bfconvert}. Skipping installation.")
             return
 
-        import textwrap
-        import subprocess
         logger.info("⚙️Installing Bio-Formats tools.")
         self.tools_dir.mkdir(parents=True, exist_ok=True)
 
@@ -44,8 +63,6 @@ class BEAT:
         What are the corresponding numbers for `2024-05-31_n98-n100`
 
         """
-        import pandas as pd
-        from ai4bmr_core.utils.tidy import tidy_name
 
         clinical = pd.read_excel(self.raw_dir / 'H_E_Images/Image_ID_vs_patient_blockIDs.xlsx')
         clinical.columns = clinical.columns.map(tidy_name)
@@ -110,9 +127,6 @@ class BEAT:
         metadata.to_parquet(save_path, engine='fastparquet')
 
     def prepare_wsi(self, force: bool = False):
-        import textwrap
-        import subprocess
-
         self.prepare_tools()
 
         dataset_dir = self.datasets_dir / 'beat'
@@ -130,13 +144,18 @@ class BEAT:
                 '1GVQ.06.2',  # barcode
             ]:
                 continue
-            
+
             intermediate_path = dataset_dir / save_name / 'intermediate.tiff'
             save_path = dataset_dir / save_name / 'wsi.tiff'
 
             if save_path.exists() and not force:
                 logger.info(f"File {save_path} already exists. Skipping conversion.")
+                continue
+            else:
+                logger.info(f'Converting {wsi_path} to {save_name}.')
+                continue
 
+            assert False
             job_name = f"proj=beat-task=wsi2tiff-{save_name}"
             logger.info(f"Converting {wsi_path} to {save_path} using Bio-Formats (job_name={job_name})")
             num_cpus = 16
@@ -181,15 +200,11 @@ class BEAT:
             subprocess.run(sbatch_command, shell=True, check=True)
 
     def post_process_tiff_flags(self):
-        from ai4bmr_learn.utils.slides import get_mpp
-        import openslide
-        import textwrap
-        import subprocess
 
         wsi_paths = list(self.datasets_dir.rglob('*wsi.tiff'))
         for wsi_path in wsi_paths:
             slide = openslide.OpenSlide(wsi_path)
-            
+
             try:
                 mpp = get_mpp(slide)
             except AssertionError:
@@ -210,9 +225,8 @@ class BEAT:
                 cmd = textwrap.dedent(cmd).strip()
                 subprocess.run(cmd, shell=True, check=True)
 
[…]
