### 2025-08-23 1a0bb0c
use ssd paths
