### 2025-08-23 5d02728
refactor: process BEAT dataset on a per-sample basis
