+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -16,8 +16,7 @@ from ai4bmr_learn.datasets.items import Patches, get_patch
 from ai4bmr_learn.plotting.contours import visualize_contours
 from ai4bmr_learn.plotting.patches import visualize_coords
 from ai4bmr_learn.utils.images import filter_coords
-from ai4bmr_learn.utils.slides import get_coordinates_dict, get_mpp, get_mpp_and_resolution, get_seg_model, \
-    get_slide_patcher_params, get_thumbnail, segment_slide
+from ai4bmr_learn.utils.slides import get_coordinates_dict, get_mpp, get_mpp_and_resolution, get_seg_model,     get_slide_patcher_params, get_thumbnail, segment_slide
 from loguru import logger
 from openslide import OpenSlide
 from skimage.io import imsave
@@ -126,7 +125,7 @@ class BEAT:
         save_path.parent.mkdir(parents=True, exist_ok=True)
         metadata.to_parquet(save_path, engine='fastparquet')
 
-    def prepare_wsi(self, force: bool = False):
+    def prepare_wsi(self, sample_id: str, force: bool = False):
         self.prepare_tools()
 
         dataset_dir = self.datasets_dir / 'beat'
@@ -135,174 +134,179 @@ class BEAT:
         clinical = pd.read_parquet(self.processed_dir / 'metadata' / 'clinical.parquet', engine='fastparquet')
         bfconvert = self.tools_dir / "bftools" / "bfconvert"
 
-        wsi_paths = list(self.base_dir.rglob('*.ndpi')) + list(self.base_dir.rglob('*.czi'))
-        for wsi_path in wsi_paths:
+        wsi_path = Path(clinical.loc[sample_id, 'wsi_path'])
+        save_name = sample_id
 
-            save_name = clinical[clinical.wsi_path == str(wsi_path)].index.item()
-            if save_name in [
-                '1GUG.0M.2',  # barcode
-                '1GVQ.06.2',  # barcode
-            ]:
-                continue
+        if save_name in [
+            '1GUG.0M.2',  # barcode
+            '1GVQ.06.2',  # barcode
+        ]:
+            logger.info(f"Skipping sample {save_name} based on barcode.")
+            return
 
-            intermediate_path = dataset_dir / save_name / 'intermediate.tiff'
-            save_path = dataset_dir / save_name / 'wsi.tiff'
+        intermediate_path = dataset_dir / save_name / 'intermediate.tiff'
+        save_path = dataset_dir / save_name / 'wsi.tiff'
 
-            if save_path.exists() and not force:
-                logger.info(f"File {save_path} already exists. Skipping conversion.")
-                continue
-            else:
-                logger.info(f'Converting {wsi_path} to {save_name}.')
-                continue
+        if save_path.exists() and not force:
+            logger.info(f"File {save_path} already exists. Skipping conversion.")
+            return
 
-            assert False
-            job_name = f"proj=beat-task=wsi2tiff-{save_name}"
-            logger.info(f"Converting {wsi_path} to {save_path} using Bio-Formats (job_name={job_name})")
-            num_cpus = 16
+        logger.info(f'Converting {wsi_path} to {save_name}.')
 
-            sbatch_command = f"""
-            sbatch \\
-            --job-name={job_name} \\
-            --output=/work/FAC/FBM/DBC/mrapsoma/prometex/logs/adrianom/{job_name}-%j.log \\
-            --error=/work/FAC/FBM/DBC/mrapsoma/prometex/logs/adrianom/{job_name}-%j.err \\
-            --time=24:00:00 \\
-            --mem=128G \\
-            --cpus-per-task={num_cpus} \\
-            <<'EOF'
-            #!/bin/bash
+        job_name = f"proj=beat-task=wsi2tiff-{save_name}"
[…]
