+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -32,8 +32,8 @@ class BEAT:
     def __init__(self, base_dir: Path | None = None):
         self.base_dir = base_dir or Path('/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat')
         self.raw_dir = self.base_dir / '01_raw'
-        self.processed_dir = self.base_dir / '02_processed'
-        self.datasets_dir = self.processed_dir / 'datasets'
+        self.processed_dir = self.base_dir / '02_processed_v2'
+        self.dataset_dir = self.processed_dir / 'datasets' / 'beat'
         self.tools_dir = self.base_dir / '03_tools'
 
     def prepare_tools(self):
@@ -126,147 +126,147 @@ class BEAT:
         save_path.parent.mkdir(parents=True, exist_ok=True)
         metadata.to_parquet(save_path, engine='fastparquet')
 
-    def prepare_wsi(self, force: bool = False):
+    def get_wsi_path(self, sample_id: str):
+        wsi_path = clinical[clinical.index == sample_id].wsi_path.item()
+        assert Path(wsi_path).exists()
+        return wsi_path
+
+    def get_clinical_metadata(self):
+        return pd.read_parquet(self.processed_dir / 'metadata' / 'clinical.parquet', engine='fastparquet')
+
+    def prepare_wsi(self, sample_id: str, force: bool = False):
+
+        if sample_id in ['1GUG.0M.2',  '1GVQ.06.2']:
+            return
+
         self.prepare_tools()
 
-        dataset_dir = self.datasets_dir / 'beat'
+        dataset_dir = self.dataset_dir
         dataset_dir.mkdir(parents=True, exist_ok=True)
 
-        clinical = pd.read_parquet(self.processed_dir / 'metadata' / 'clinical.parquet', engine='fastparquet')
+        clinical = self.get_clinical_metadata()
         bfconvert = self.tools_dir / "bftools" / "bfconvert"
 
-        wsi_paths = list(self.base_dir.rglob('*.ndpi')) + list(self.base_dir.rglob('*.czi'))
-        for wsi_path in wsi_paths:
+        wsi_path = self.get_wsi_path(sample_id)
 
-            save_name = clinical[clinical.wsi_path == str(wsi_path)].index.item()
-            if save_name in [
-                '1GUG.0M.2',  # barcode
-                '1GVQ.06.2',  # barcode
-            ]:
-                continue
+        intermediate_path = dataset_dir / sample_id / 'intermediate.tiff'
+        save_path = dataset_dir / sample_id / 'wsi.tiff'
 
-            intermediate_path = dataset_dir / save_name / 'intermediate.tiff'
-            save_path = dataset_dir / save_name / 'wsi.tiff'
+        if save_path.exists() and not force:
+            logger.info(f"File {save_path} already exists. Skipping conversion.")
+            return
+        else:
+            logger.info(f'Converting {wsi_path} to {sample_id}.')
+            return
 
-            if save_path.exists() and not force:
-                logger.info(f"File {save_path} already exists. Skipping conversion.")
-                continue
-            else:
-                logger.info(f'Converting {wsi_path} to {save_name}.')
-                continue
+        assert False
+        job_name = f"{sample_id}-wsi2tiff"
+        logger.info(f"Converting {wsi_path} to {save_path} using Bio-Formats (job_name={job_name})")
+        num_cpus = 16
 
-            assert False
-            job_name = f"proj=beat-task=wsi2tiff-{save_name}"
-            logger.info(f"Converting {wsi_path} to {save_path} using Bio-Formats (job_name={job_name})")
-            num_cpus = 16
+        sbatch_command = f"""
[…]
