+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -1,8 +1,7 @@
-from pathlib import Path
+import json
 import json
 import subprocess
 import textwrap
-from dataclasses import asdict
 from pathlib import Path
 
 import geopandas as gpd
@@ -12,8 +11,7 @@ import pandas as pd
 import torch
 from ai4bmr_core.utils.tidy import tidy_name
 from ai4bmr_learn.data_models.Coordinate import SlideCoordinate
-from ai4bmr_learn.datasets.items import Patches, get_patch
-from ai4bmr_learn.plotting.contours import visualize_contours
+from ai4bmr_learn.datasets.items import SlidePatches, get_patch
 from ai4bmr_learn.plotting.patches import visualize_coords
 from ai4bmr_learn.utils.images import filter_coords
 from ai4bmr_learn.utils.slides import get_coordinates_dict, get_mpp, get_mpp_and_resolution, get_seg_model, \
@@ -22,9 +20,9 @@ from loguru import logger
 from openslide import OpenSlide
 from skimage.io import imsave
 from torch.utils.data import DataLoader
-from torchvision.transforms import v2
+from torchvision.transforms import InterpolationMode, v2
 from tqdm import tqdm
-from trident.patch_encoder_models import encoder_factory, encoder_factory
+from trident.patch_encoder_models import encoder_factory
 
 
 class BEAT:
@@ -32,7 +30,7 @@ class BEAT:
     def __init__(self, base_dir: Path | None = None):
         self.base_dir = base_dir or Path('/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat')
         self.raw_dir = self.base_dir / '01_raw'
-        self.processed_dir = self.base_dir / '02_processed_v2'
+        self.processed_dir = self.base_dir / '02_processed'
         self.dataset_dir = self.processed_dir / 'datasets' / 'beat'
         self.tools_dir = self.base_dir / '03_tools'
 
@@ -127,6 +125,7 @@ class BEAT:
         metadata.to_parquet(save_path, engine='fastparquet')
 
     def get_wsi_path(self, sample_id: str):
+        clinical = self.get_clinical_metadata()
         wsi_path = clinical[clinical.index == sample_id].wsi_path.item()
         assert Path(wsi_path).exists()
         return wsi_path
@@ -205,7 +204,7 @@ class BEAT:
 
     def post_process_tiff_flags(self, sample_id: str):
 
-        wsi_paths = self.get_wsi_path(sample_id)
+        wsi_path = self.get_wsi_path(sample_id)
         slide = openslide.OpenSlide(wsi_path)
 
         try:
@@ -242,21 +241,20 @@ class BEAT:
         save_contours_path = sample_dir / 'contours' / f'{version_name}.parquet'
 
         if save_contours_path.exists() and save_contours_path.exists():
-            logger.info(f'[{i}/{len(sample_dirs)}] {sample_dir} is already segmented')
+            logger.info(f'{sample_id} is already segmented')
             return
 
-        logger.info(f'[{i}/{len(sample_dirs)}] 🧩 segmenting {sample_dir}')
+        logger.info(f'🧩 segmenting {sample_id}')
 
         slide = OpenSlide(img_path)
 
         try:
             mpp = get_mpp_and_resolution(slide)[0]
         except AssertionError:
-            logger.error(f'Error retriving slide mpp for {img_path}')
+            logger.error(f'Error retrieving slide mpp for {img_path}')
             return
 
-        logger.info(
-            f'{sample_dir.name} reports mpp={mpp:.4f}' + (f', assuming mpp={source_mpp}' if source_mpp else ''))
+        logger.info(f'{sample_dir.name} reports mpp={mpp:.4f}')
 
[…]
