+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -1,5 +1,4 @@
 import json
-import json
 import subprocess
 import textwrap
 from pathlib import Path
@@ -30,9 +29,13 @@ class BEAT:
     def __init__(self, base_dir: Path | None = None):
         self.base_dir = base_dir or Path('/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat')
         self.raw_dir = self.base_dir / '01_raw'
-        self.processed_dir = self.base_dir / '02_processed'
+        self.processed_dir = self.base_dir / '02_processed_v2'
         self.dataset_dir = self.processed_dir / 'datasets' / 'beat'
         self.tools_dir = self.base_dir / '03_tools'
+
+        # GLOBAL PARAMS
+        self.target_mpp = 0.8624999831542972
+        self.overlap = 0.25
 
     def prepare_tools(self):
         bfconvert = self.tools_dir / "bftools" / "bfconvert"
@@ -124,14 +127,14 @@ class BEAT:
         save_path.parent.mkdir(parents=True, exist_ok=True)
         metadata.to_parquet(save_path, engine='fastparquet')
 
+    def get_clinical_metadata(self):
+        return pd.read_parquet(self.processed_dir / 'metadata' / 'clinical.parquet', engine='fastparquet')
+
     def get_wsi_path(self, sample_id: str):
         clinical = self.get_clinical_metadata()
         wsi_path = clinical[clinical.index == sample_id].wsi_path.item()
         assert Path(wsi_path).exists()
         return wsi_path
-
-    def get_clinical_metadata(self):
-        return pd.read_parquet(self.processed_dir / 'metadata' / 'clinical.parquet', engine='fastparquet')
 
     def prepare_wsi(self, sample_id: str, force: bool = False):
 
@@ -203,8 +206,9 @@ class BEAT:
         subprocess.run(sbatch_command, shell=True, check=True)
 
     def post_process_tiff_flags(self, sample_id: str):
+        logger.info(f'Setting tiff flags for {sample_id}')
 
-        wsi_path = self.get_wsi_path(sample_id)
+        wsi_path = self.dataset_dir / sample_id / 'wsi.tiff'
         slide = openslide.OpenSlide(wsi_path)
 
         try:
@@ -228,6 +232,7 @@ class BEAT:
             subprocess.run(cmd, shell=True, check=True)
 
     def segment(self, sample_id: str, model_name: str = 'grandqc', target_mpp: float = 4.0):
+        logger.info(f'🧩 Segmenting {sample_id} with {model_name}')
 
         seg_model = get_seg_model(model_name=model_name)  # hest, grandqc, grandqc_artifact
         sample_dir = self.dataset_dir / sample_id
@@ -243,8 +248,6 @@ class BEAT:
         if save_contours_path.exists() and save_contours_path.exists():
             logger.info(f'{sample_id} is already segmented')
             return
-
-        logger.info(f'🧩 segmenting {sample_id}')
 
         slide = OpenSlide(img_path)
 
@@ -280,25 +283,6 @@ class BEAT:
         save_path = img_path.parent / 'thumbnail.png'
         imsave(str(save_path), thumbnail)
 
-    def prepare_data(self, force: bool = False):
-        self.prepare_tools()
-        self.prepare_clinical()
-
-        self.prepare_wsi(force=force)
-        self.post_process_tiff_flags()
-
-        self.segment(model_name='hest', target_mpp=4)
-        self.segment(model_name='grandqc', target_mpp=4)
-
-        self.create_coords(patch_size=224)
-        self.create_coords(patch_size=512)
-
-        patch_size = patch_stride = 512
-        target_mpp = 0.8625
-        overlap = 0.25
[…]
