### 2025-08-23 103c3fe
add: get_sample_ids
