### 2025-08-23 1d6f172
remove assert
