### 2025-08-23 30f0d4e
add env and profile, remove samples-id and load them with the BEAT class directlu
