+++ workflows/BEAT/sample_ids.txt
@@ -1,183 +1,0 @@
-1HS7.06.0
-1IM0.0A.0
-1IVR.2T.0
-1H5M.0H.0
-1IM1.0A.0
-1HUX.06.0
-1H3W.06.0
-1H49.0M.0
-1JAR.07.0
-1HYU.06.0
-1J7Y.06.0
-1J6H.0G.0
-1IVR2V.0
-1H5L.06.0
-1H49.0K.0
-1IM0.0C.0
-1IQJ.06.0
-1H49.0I.0
-1IWW.05.0
-1IK0.06.0
-1J25.05.0
-1IZY.05.0
-1IUV.05.0
-1I31.06.0
-1H4E.06.0
-1IVR.2Z.0
-1J4J.1C.0
-1H49.0O.0
-1IJZ.06.0
-1IS2.0E.0
-1JAS.07.0
-1H5O.06.0
-1HLN.06.0
-1HV3.06.0
-1JAT.07.0
-1J1D.0L.0
-1IS3.06.0
-1IR5.06.0
-1IXC.05.0
-1IQK.06.0
-1H5K.0Q.0
-1HS6.0C.0
-1J4B.05.0
-1IRS.10.0
-1I32.06.0
-1IO9.0B.0
-1IUW.05.0
-1J30.05.0
-1ISE.06.0
-1IS2.0G.0
-1IPZ.06.0
-1J4I.05.0
-1HS6.0A.0
-1IVQ.1Y.0
-1J00.05.0
-1IZX.05.0
-1I30.06.0
-1IVR.2X.0
-1I33.06.0
-1IUK.05.0
-1IVQ.1V.0
-1H5K.0M.0
-1IKT.0X.0
-1H5K.0O.0
-1J2Y.05.0
-1H4D.06.0
-1J2X.06.0
-1IJ5.0G.0
-1J31.05.0
-1GAY.06.0
-1G50.06.0
-1G40.06.0
-1G3T.06.0
-1G11.0G.0
-1G9B.06.0
-1G3S.06.0
-1G63.06.0
-1G12.07.0
-1GJB.06.0
-1G3X.06.0
-1G4Z.06.0
-1G6Z.06.0
-1G14.06.0
-1G5A.0C.0
-1G59.06.0
-1G51.06.0
-1G4J.0C.0
-1G3Z.06.0
-1G3Y.0E.0
-1G5B.06.0
-1G52.06.0
-1G4K.0F.0
-1G3U.06.0
-1G9A.06.0
-1FZX.0I.0
-1G4X.06.0
-1G3W.06.0
-1G5A.0A.0
-1GNX.06.0
-1G11.0E.0
-1G11.0I.0
-1FYV.06.0
-1FYO.06.0
-1FU4.06.0
-1FXS.06.0
-1FYS.06.0
-1FU2.06.0
-1FYN.06.0
-1FP2.0E.0
-1FYQ.06.0
-1FU3.06.0
-1FYT.06.0
-1FYP.06.0
-1GPT.06.0
-1GGR.06.0
-1GNV.06.0
-1GPZ.06.0
-1GTL.06.0
-1GCZ.0D.0
-1GQ8.06.0
-1GS6.0H.0
-1GTD.06.0
-1GPO.06.0
-1GPP.06.0
-1GS0.06.0
-1GEP.06.0
-1GJC.06.0
-1GWV.06.0
-1GIL.0I.0
-1GCY.0G.0
-1GK0.06.0
-1GEG.0G.0
-1GTE.06.0
-1GTK.12.0
-1GVX.0H.0
-1GP1.06.0
-1GEN.06.0
-1GGF.06.0
[…]
