### 2025-08-23 db99982
update env with vips and housekeep
