### 2025-08-23 2f91a85
set mpp of slides with 0.08 to 0.8, refine Snakemake
