### 2025-08-25 ef6a21c
update snakemake pipeline and runt trident workflow
