### 2025-08-25 b553d81
update trident pipeline
