### 2025-08-25 0ab838c
update trident pipeline
