+++ workflows/BEAT/envs/omics.yaml
@@ -1,0 +1,409 @@
+name: omics
+channels:
+  - bioconda
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1
+  - _openmp_mutex=4.5
+  - adwaita-icon-theme=48.1
+  - amply=0.1.6
+  - annotated-types=0.7.0
+  - aom=3.9.1
+  - appdirs=1.4.4
+  - argparse-dataclass=2.0.0
+  - at-spi2-atk=2.38.0
+  - at-spi2-core=2.40.3
+  - atk-1.0=2.38.0
+  - attrs=25.3.0
+  - brotli-python=1.1.0
+  - bzip2=1.0.8
+  - c-ares=1.34.5
+  - ca-certificates=2025.8.3
+  - cairo=1.18.4
+  - certifi=2025.8.3
+  - cffi=1.17.1
+  - cfitsio=4.6.2
+  - charset-normalizer=3.4.3
+  - click=8.2.1
+  - coin-or-cbc=2.10.12
+  - coin-or-cgl=0.60.9
+  - coin-or-clp=1.17.10
+  - coin-or-osi=0.108.11
+  - coin-or-utils=2.11.12
+  - colorama=0.4.6
+  - coloredlogs=15.0.1
+  - conda-inject=1.3.2
+  - configargparse=1.7.1
+  - connection_pool=0.0.3
+  - dav1d=1.2.1
+  - dbus=1.16.2
+  - docutils=0.22
+  - dpath=2.2.0
+  - eido=0.2.4
+  - epoxy=1.5.10
+  - exceptiongroup=1.3.0
+  - fftw=3.3.10
+  - font-ttf-dejavu-sans-mono=2.37
+  - font-ttf-inconsolata=3.000
+  - font-ttf-source-code-pro=2.038
+  - font-ttf-ubuntu=0.83
+  - fontconfig=2.15.0
+  - fonts-conda-ecosystem=1
+  - fonts-conda-forge=1
+  - freetype=2.13.3
+  - fribidi=1.0.10
+  - gdk-pixbuf=2.42.12
+  - ghostscript=10.04.0
+  - giflib=5.2.2
+  - gitdb=4.0.12
+  - gitpython=3.1.45
+  - glib-tools=2.84.3
+  - graphite2=1.3.14
+  - graphviz=13.1.2
+  - gtk3=3.24.43
+  - gts=0.7.6
+  - h2=4.2.0
+  - harfbuzz=11.4.2
+  - hdf5=1.14.6
+  - hicolor-icon-theme=0.17
+  - hpack=4.1.0
+  - humanfriendly=10.0
+  - hyperframe=6.1.0
+  - icu=75.1
+  - idna=3.10
+  - imagemagick=7.1.2_1
+  - immutables=0.21
+  - iniconfig=2.0.0
+  - jbig=2.1
+  - jinja2=3.1.6
+  - jsonschema=4.25.1
+  - jsonschema-specifications=2025.4.1
+  - jupyter_core=5.8.1
+  - keyutils=1.6.3
+  - krb5=1.21.3
+  - lcms2=2.17
+  - ld_impl_linux-64=2.44
+  - lerc=4.0.0
+  - libaec=1.1.4
+  - libarchive=3.8.1
+  - libavif16=1.3.0
+  - libblas=3.9.0
+  - libcblas=3.9.0
+  - libcups=2.3.3
+  - libcurl=8.14.1
+  - libde265=1.0.15
+  - libdeflate=1.24
+  - libdicom=1.2.0
+  - libedit=3.1.20250104
+  - libev=4.33
+  - libexif=0.6.21
[…]
