### 2025-08-25 5350b4c
run trident and snakemake
