+++ tests/datasets/test_BEAT.py
@@ -2,6 +2,6 @@ def test_BEAT():
     from ai4bmr_datasets import BEAT
     ds = self = BEAT()
     ds.prepare_tools()
-    ds.prepare_clinical()
+    ds.prepare_metadata()
     ds.prepare_wsi(force=True)
     # ds.setup()
