+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -7,6 +7,7 @@ import geopandas as gpd
 import numpy as np
 import openslide
 import pandas as pd
+import tifffile
 import torch
 from ai4bmr_core.utils.tidy import tidy_name
 from ai4bmr_learn.data_models.Coordinate import SlideCoordinate
@@ -24,15 +25,6 @@ from torchvision.transforms import InterpolationMode, v2
 from tqdm import tqdm
 from trident.patch_encoder_models import encoder_factory
 from ai4bmr_learn.plotting.umap import plot_umap
-
-
-# NOTES: check segmentation of
-# - 1FU3.06.0
-# - 1HRW.06.0 reports mpp=0.0863
-# - 1G59.06.0 reports mpp=0.0863
-# - 1G3S.06.0 reports mpp=0.0863
-# - 1GP0.06.0 reports mpp=0.0863
-
 
 class BEAT:
     INVALID_SAMPLE_IDS = {
@@ -73,7 +65,7 @@ class BEAT:
         install_cmd = textwrap.dedent(install_cmd).strip()
         subprocess.run(install_cmd, shell=True, check=True)
 
-    def prepare_clinical(self):
+    def prepare_metadata(self):
         """Questions
         Is this n=12 or n=14? I assume -14 can be ignored
         2024-05-24_n.1 - n.14/malexan3-24-05-2024-012-14.czi
@@ -140,20 +132,20 @@ class BEAT:
         metadata = metadata.convert_dtypes().astype({'wsi_path': str})
         metadata = metadata.set_index('sample_id')
 
-        save_path = self.processed_dir / 'metadata' / 'clinical.parquet'
+        save_path = self.dataset_dir / 'metadata.parquet'
         save_path.parent.mkdir(parents=True, exist_ok=True)
         metadata.to_parquet(save_path, engine='fastparquet')
 
-    def get_clinical_metadata(self):
-        return pd.read_parquet(self.processed_dir / 'metadata' / 'clinical.parquet', engine='fastparquet')
+    def get_metadata(self):
+        return pd.read_parquet(self.dataset_dir / 'metadata.parquet', engine='fastparquet')
 
     def get_sample_ids(self):
-        clinical = self.get_clinical_metadata()
+        clinical = self.get_metadata()
         sample_ids = set(clinical.index.tolist()) - self.INVALID_SAMPLE_IDS
         return sample_ids
 
     def get_wsi_path(self, sample_id: str):
-        clinical = self.get_clinical_metadata()
+        clinical = self.get_metadata()
         wsi_path = clinical[clinical.index == sample_id].wsi_path.item()
         assert Path(wsi_path).exists()
         return wsi_path
@@ -515,7 +507,7 @@ class BEAT:
         ax.figure.savefig(save_path, dpi=300)
 
 
-    def create_umap(self, sample_id: str, model_name: str, coords_version: str):
+    def create_wsi_umap(self, sample_id: str, model_name: str, coords_version: str):
         logger.info(f'Computing UMAP for {sample_id}.')
 
         embeddings_dir = self.dataset_dir / sample_id / 'embeddings' / 'patch' / f'model={model_name}' / coords_version
@@ -534,7 +526,7 @@ class BEAT:
         ax.figure.tight_layout()
         ax.figure.savefig(save_path, dpi=300)
 
-    def create_slides_umaps(self):
+    def create_slides_umap(self):
 
         embeddings_dir: Path = Path(
             '/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat/02_processed_v2/datasets/wsi_embed_subset/20x_512px_0px_overlap/slide_features_titan')
@@ -549,29 +541,39 @@ class BEAT:
[…]
