### 2025-08-26 6228764
run snakemake on BEAT
