+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -1,161 +1,0 @@
-import shutil
-from pathlib import Path
-
-import pandas as pd
-from loguru import logger
-
-class BEAT:
-
-    def __init__(self, base_dir: Path | None = None):
-        self.base_dir = base_dir or Path('/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat')
-        self.raw_dir = self.base_dir / '01_raw'
-        self.processed_dir = self.base_dir / '02_processed'
-        self.tools_dir = self.base_dir / '03_tools'
-        self.log_dir = self.base_dir / '04_logs'
-
-        self.images_dir = self.processed_dir / 'images'
-        self.images_dir.mkdir(parents=True, exist_ok=True)
-
-    def prepare_tools(self):
-        bfconvert = self.tools_dir / "bftools" /"bfconvert"
-
-        if bfconvert.exists():
-            logger.info(f"bfconvert found at {bfconvert}. Skipping installation.")
-            return
-
-        import textwrap
-        import subprocess
-        logger.info("⚙️Installing Bio-Formats tools.")
-        self.tools_dir.mkdir(parents=True, exist_ok=True)
-
-        install_cmd = f"""
-        wget https://downloads.openmicroscopy.org/bio-formats/8.2.0/artifacts/bftools.zip -O {self.tools_dir}/bftools.zip
-        unzip {self.tools_dir}/bftools.zip
-        {self.tools_dir / "bftools" /"bfconvert"} -version || exit
-        """
-
-        install_cmd = textwrap.dedent(install_cmd).strip()
-        subprocess.run(install_cmd, shell=True, check=True)
-
-    def prepare_clinical(self):
-        """Questions
-        Is this n=12 or n=14? I assume -14 can be ignored
-        2024-05-24_n.1 - n.14/malexan3-24-05-2024-012-14.czi
-
-        What are the corresponding numbers for `2024-05-31_n98-n100`
-
-        """
-        import pandas as pd
-        from ai4bmr_core.utils.tidy import tidy_name
-
-        clinical = pd.read_excel(self.raw_dir / 'H_E_Images/Image_ID_vs_patient_blockIDs.xlsx')
-        clinical.columns = clinical.columns.map(tidy_name)
-        clinical = clinical.convert_dtypes()
-
-        assert clinical.sample_barcode.value_counts().max() == 1
-
-        dir_to_range = {
-            '2024-05-28 slide n.15 - n.50': list(range(15, 50 + 1)),
-            'Slide N°130-222': list(range(130, 222 + 1)),
-            '2024-05-24_n.1 - n.14': list(range(1, 14 + 1)),
-            '2024-05-30_n51 - n.97': list(range(51, 97 + 1)),
-            '2024-06-05_n101-n129': list(range(101, 129 + 1)),
-            '2024-05-31_n98-n100': list(range(98, 100 + 1))
-        }
-
-        import re
-        wsi_paths = list(self.base_dir.rglob('*.ndpi')) + list(self.base_dir.rglob('*.czi'))
-        records = []
-        for wsi_path in wsi_paths:
-
-            match = re.search(r'malexan3-31-05-2024-(\d{3})', str(wsi_path))
-            if match is not None:
-                index = int(match.group(1)) - 60
-                n = dir_to_range[wsi_path.parent.name][index]
-                records.append({'wsi_path': wsi_path, 'n': n})
-                continue
-
-            match = re.search(r'malexan3-\d{2}-\d{2}-\d{4}-(\d{3})\.czi$', str(wsi_path))
-            if match is not None:
-                index = int(match.group(1))
-                if wsi_path.parent.name == '2024-06-05_n101-n129':
-                    index -= 31
[…]
