+++ src/ai4bmr_datasets/datasets/PCa.py
@@ -1,1088 +1,0 @@
-from loguru import logger
-from ai4bmr_core.utils.tidy import tidy_name
-from tifffile import imread, imwrite
-from pathlib import Path
-import pandas as pd
-from ai4bmr_datasets.datasets.BaseIMCDataset import BaseIMCDataset
-
-
-class PCa(BaseIMCDataset):
-    name = "PCa"
-    id = "PCa"
-    doi = "unpublished"
-
-    def __init__(self,
-                 base_dir: Path | None = None,
-                 image_version: str | None = None,
-                 mask_version: str | None = None,
-                 feature_version: str | None = None,
-                 metadata_version: str | None = None,
-                 load_intensity: bool = False,
-                 load_spatial: bool = False,
-                 load_metadata: bool = False,
-                 align: bool = False,
-                 join: str = "outer",
-                 ):
-        """
-        Initializes the Cords2024 dataset.
-
-        Args:
-            base_dir (Path | None): Base directory for the dataset.
-            image_version (str | None): Version of the images to load.
-            mask_version (str | None): Version of the masks to load.
-            feature_version (str | None): Version of the features to load.
-            metadata_version (str | None): Version of the metadata to load.
-            load_intensity (bool): Whether to load intensity features.
-            load_spatial (bool): Whether to load spatial features.
-            load_metadata (bool): Whether to load metadata.
-            align (bool): Whether to align sample IDs across modalities.
-            join (str): Type of join for alignment (e.g., 'outer').
-        """
-
-        super().__init__(base_dir=base_dir,
-                         image_version=image_version,
-                         mask_version=mask_version,
-                         feature_version=feature_version,
-                         metadata_version=metadata_version,
-                         load_intensity=load_intensity,
-                         load_spatial=load_spatial,
-                         load_metadata=load_metadata,
-                         align=align,
-                         join=join)
-
-        # raw paths
-        self.raw_clinical_metadata_path = (
-                self.base_dir / "01_raw" / "metadata" / "ROI_matching_blockID.xlsx"
-        )
-        self.raw_tma_annotations_path = (
-                self.base_dir / "01_raw" / "metadata" / "tma-annotations-v2.xlsx"
-        )
-
-        self.raw_tma_tidy_path = self.base_dir / "01_raw" / "metadata" / "TMA_tidy.txt"
-        self.raw_annotations_path = self.base_dir / "01_raw" / "annotations" / "labels.parquet"
-
-    def prepare_data(self):
-        self.process_acquisitions()
-        self.create_panel()
-
-        self.create_images()
-        self.create_filtered_images()
-
-        self.create_masks()
-        self.create_filtered_masks()
-
-        self.create_clinical_metadata()
-        self.create_tma_annotations()
-
-        self.label_transfer()
-        self.create_annotated()
-        self.compute_features(image_version="filtered", mask_version="annotated")
-
-    def process_acquisitions(self):
-        """
-        Failed to read acquisitions:
-          - 240210_IIIBL_X3Y15_150_1
-        """
-        from readimc import MCDFile, TXTFile
-        from tifffile import imwrite
-        import json
-
-        acquisitions_dir = self.raw_dir / 'acquisitions'
-        acquisitions_dir.mkdir(parents=True, exist_ok=True)
-
-        num_failed_reads = 0
-
-        mcd_paths = [i for i in (self.raw_dir / 'raw').rglob("*.mcd") if i.is_file()]
-        for mcd_path in mcd_paths:
-            logger.info(f"Processing {mcd_path.name}")
-            slide_date = mcd_path.stem
-
-            with MCDFile(mcd_path) as f:
[…]
