### 2025-08-27 c90ddff
drop ai4bmr_core dependency and torch
