### 2025-08-27 5ca57b4
add imagecodescs and zarr
