### 2025-08-27 ae1fce1
upate trident params and fix merge on PCa
