+++ src/ai4bmr_datasets/datasets/PCa.py
@@ -649,7 +649,7 @@ class PCa(BaseIMCDataset):
         )
 
         mapping = self.get_napari_file_name_to_sample_id_mapping()
-        sample_ids = roi_match_blockId.file_name_napari.map(mapping).to_list()
+        sample_ids = roi_match_blockId.file_name_napari.map(mapping)
 
         roi_match_blockId = roi_match_blockId.assign(
             sample_id=sample_ids,
@@ -658,14 +658,12 @@ class PCa(BaseIMCDataset):
             tma_coordinates=tma_coordinates,
         )
 
-        roi_match_blockId.loc[
-            roi_match_blockId.tma_sample_id == "150 - split", "tma_sample_id"
-        ] = "150"
+        # NOTE: we drop the corrupted acquisition, and simplify the name of the rescued split
+        roi_match_blockId = roi_match_blockId[roi_match_blockId.sample_id.notna()]  # this is tma_id "150"
+        roi_match_blockId.loc[roi_match_blockId.tma_sample_id == "150 - split", "tma_sample_id"] = "150"
         assert roi_match_blockId.tma_sample_id.str.contains("split").sum() == 0
 
-        patient_ids = (
-            roi_match_blockId["Original block number"].str.split("_").str[0].str.strip()
-        )
+        patient_ids = roi_match_blockId["Original block number"].str.split("_").str[0].str.strip()
         roi_match_blockId = roi_match_blockId.assign(PAT_ID=patient_ids)
 
         # NOTE: these are the LP samples (i.e. test runs)
@@ -675,8 +673,7 @@ class PCa(BaseIMCDataset):
         assert not roi_match_blockId.slide_code.isna().any()
         assert not roi_match_blockId.tma_coordinates.isna().any()
         assert not roi_match_blockId.tma_sample_id.isna().any()
-        # NOTE: this is the corrupted acquisition, IIIBL_X3Y15_150  240210_001.tiff
-        assert roi_match_blockId.sample_id.isna().sum() == 1
+        assert roi_match_blockId.sample_id.isna().sum() == 0
 
         tma_id = roi_match_blockId.slide_code + "_" + roi_match_blockId.tma_coordinates
         roi_match_blockId["tma_id"] = tma_id
@@ -739,7 +736,6 @@ class PCa(BaseIMCDataset):
         ]
 
         metadata["date_of_surgery"] = pd.to_datetime(metadata["date_of_surgery"])
-
         metadata["age_at_surgery"] = metadata["age_at_surgery"].astype(int)
 
         mapping = {
@@ -748,9 +744,7 @@ class PCa(BaseIMCDataset):
             "Only 1 Gleason pattern": "only_1",
             "unkown": "unknown",
         }
-        metadata["gleason_pattern_tma_core"] = metadata.gleason_pattern_tma_core.map(
-            mapping
-        )
+        metadata["gleason_pattern_tma_core"] = metadata.gleason_pattern_tma_core.map(mapping)
 
         metadata["psa_at_surgery"] = metadata.psa_at_surgery.astype(float)
         assert metadata["psa_at_surgery"].isna().sum() == 0
@@ -758,9 +752,7 @@ class PCa(BaseIMCDataset):
         metadata["last_fu"] = metadata["last_fu"].astype(int)
         assert metadata.last_fu.isna().sum() == 0
 
-        metadata["cause_of_death"] = metadata.cause_of_death.map(
-            {"0": "alive", "1": "non-PCa_death", "2": "PCa_death"}
-        )
+        metadata["cause_of_death"] = metadata.cause_of_death.map({"0": "alive", "1": "non-PCa_death", "2": "PCa_death"})
         assert metadata.cause_of_death.isna().sum() == 0
 
[…]
