### 2025-08-27 0651a4c
exclude hidden files
