+++ tests/datasets/test_prepare_and_setup.py
@@ -46,4 +46,4 @@ def test_prepare_data(dataset_name):
 # test_prepare_data("Jackson2020")
 # test_prepare_data("Danenberg2022")
 # test_prepare_data("Cords2024")
-# test_prepare_data("Keren2018")
+test_prepare_data("Keren2018")
