+++ docs/requirements.txt
@@ -1,3 +1,0 @@
-Sphinx
-sphinx-rtd-theme
-sphinx-autodoc-typehints
