+++ docs/source/index.rst
@@ -1,22 +1,0 @@
-.. AI4BMR Datasets documentation master file, created by
-   sphinx-quickstart on Fri Aug  8 17:04:24 2025.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-AI4BMR Datasets documentation
-=============================
-
-Add your content using ``reStructuredText`` syntax. See the
-`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
-documentation for details.
-
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-.. automodule:: ai4bmr_datasets
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
