+++ tests/datasets/test_BEAT.py
@@ -1,5 +1,0 @@
-def test_Keren2018():
-    from ai4bmr_datasets import BEAT
-    ds = self = BEAT()
-    ds.prepare_data()
-    ds.setup()
