+++ tests/datasets/test_PCa.py
@@ -1,8 +1,0 @@
-def test_PCa():
-    from ai4bmr_datasets import PCa
-
-    ds = dataset = self = PCa(image_version="filtered", mask_version="annotated",
-                              load_metadata=True, load_intensity=True, align=True)
-    # ds.prepare_data()
-    ds.setup()
-    ds.images.keys()
