### 2025-08-27 a7dbb14
propagate API changes
