+++ README.md
@@ -113,50 +113,49 @@ pip install git+https://github.com/AI4SCR/ai4bmr-datasets.git@v1.0.0
 All datasets are initialized by passing the desired versions and data loading flags to the constructor.
 
 ```python
-from ai4bmr_datasets import Jackson2020
+from ai4bmr_datasets import Keren2018
 from pathlib import Path
 
 base_dir = Path("/path/to/storage")  # can be None, resolves to ~/.cache/ai4bmr_datasets by default
-dataset = Jackson2020(
-    base_dir=None,
+dataset = Keren2018(
+    base_dir=base_dir,
     image_version="published",
     mask_version="published",
-    # Optionally, load intensity features and metadata
+    # optionally, load intensity features and metadata
     feature_version="published", 
     load_intensity=True,
     metadata_version="published", 
     load_metadata=True
 )
 dataset.prepare_data()  # Downloads and preprocesses data if needed, only needs to be run once
+dataset.setup()  # load the data
 ```
 
 ### 2. Access core components
 
 ```python
-print(dataset.sample_ids)     # List of sample IDs
-print(dataset.images.keys()) # Dictionary of images
+print(dataset.sample_ids)  # List of sample IDs
+print(dataset.images.keys())  # Dictionary of images
 print(dataset.masks.keys())  # Dictionary of masks
 ```
 
-### 3. Access optional cell-level features and metadata
+### 3. Work with image and mask objects
 
+```python
+sample_id = dataset.sample_ids[0]  # get the first sample ID
+img = dataset.images[sample_id].data
+print("Image shape:", img.shape)
+
+mask = dataset.masks[sample_id].data
+print("Mask shape:", mask.shape)
+```
+
+### 4. Access optional cell-level features and metadata
 
 ```python
 print(dataset.intensity.shape)  # Cell x marker matrix
 print(dataset.metadata.shape)   # Cell x annotation matrix
 ```
-
-### 4. Work with image and mask objects
-
-```python
-sample_id = dataset.sample_ids[0]  # get the first sample ID
-img = dataset.images[sample_id].data
-mask = dataset.masks[sample_id].data
-
-print("Image shape:", img.shape)
-print("Mask shape:", mask.shape)
-```
-
 ---
 
 ## 🤝 Contributing
