### 2025-08-27 6fd33ba
remove github workflow for docs
