### 2025-08-27 e3c6165
update README with latest release
