### 2025-08-27 f324be3
update README with latest release
