### 2025-08-27 ce689bf
update README with latest release
