### 2025-08-27 89d9ead
trigger PDF generation
