### 2025-08-27 5b31125
fix Cords paths
