+++ src/ai4bmr_datasets/datasets/Cords2024.py
@@ -3,12 +3,13 @@ import re
 import shutil
 from pathlib import Path
 
+import numpy as np
 import pandas as pd
 from loguru import logger
-from tifffile import imread, imwrite
 from tqdm import tqdm
 
 from ai4bmr_datasets.datasets.BaseIMCDataset import BaseIMCDataset
+from ai4bmr_datasets.utils import io
 from ai4bmr_datasets.utils.download import download_file_map, unzip_recursive
 from ai4bmr_datasets.utils.tidy import filter_paths
 
@@ -88,6 +89,7 @@ class Cords2024(BaseIMCDataset):
         self.create_images()
         self.create_masks()
         self.create_metadata()
+        self.create_features()
 
         self.create_annotated()
 
@@ -208,9 +210,7 @@ class Cords2024(BaseIMCDataset):
                             metal_names=acq.channel_metals,
                         )
 
-                        image_name = (
-                            f"{mcd_id}_{acq.id}"  # should be equivalent to 'Tma_ac'
-                        )
+                        image_name = f"{mcd_id}_{acq.id}"  # should be equivalent to 'Tma_ac'
                         metadata_path = acquisitions_dir / f"{image_name}.json"
                         image_path = acquisitions_dir / f"{image_name}.tiff"
 
@@ -221,7 +221,8 @@ class Cords2024(BaseIMCDataset):
                         try:
                             logger.info(f"Reading acquisition of image {image_name}")
                             img = f.read_acquisition(acq)
-                            imwrite(image_path, img)
+                            img = img.astype(np.float32)
+                            io.imsave(save_path=image_path, img=img)
                             with open(metadata_path, "w") as f_json:
                                 json.dump(item, f_json)
                         except OSError:
@@ -229,7 +230,8 @@ class Cords2024(BaseIMCDataset):
                             num_failed_reads += 1
                             continue
 
-    def validate_masks(self):
+    @classmethod
+    def validate_masks(cls):
         """
         Validates the consistency between mask objects and metadata objects.
 
@@ -239,11 +241,11 @@ class Cords2024(BaseIMCDataset):
         """
         import numpy as np
         version_name = 'published'
-        self.setup(image_version=version_name, mask_version=version_name, metadata_version=version_name,
-                   load_metadata=True, load_spatial=False, load_intensity=False)
+        ds = cls(image_version=version_name, mask_version=version_name, metadata_version=version_name,
+                 load_metadata=True, load_spatial=False, load_intensity=False)
 
         mismatches = []
-        for sample_id, mask in self.masks.items():
+        for sample_id, mask in ds.masks.items():
             uniq_mask_objs = set(map(int, np.unique(mask.data))) - {0}
             meta = self.metadata.loc[sample_id]
             uniq_meta_objs = set(meta.index.astype(int))
@@ -274,7 +276,8 @@ class Cords2024(BaseIMCDataset):
         import pandas as pd
         panel = None
 
-        for img_metadata_path in self.raw_acquisitions_dir.glob("*.json"):
+        img_metadata_paths = [p for p in self.raw_acquisitions_dir.glob("*.json") if not p.name.startswith('.')]
+        for img_metadata_path in img_metadata_paths:
             with open(img_metadata_path, "r") as f:
                 img_metadata = json.load(f)
                 df = pd.DataFrame(img_metadata)
@@ -323,21 +326,22 @@ class Cords2024(BaseIMCDataset):
         panel = pd.read_parquet(panel_path)
 
         acquisitions_dir = self.raw_acquisitions_dir
[…]
