### 2025-08-28 b0e58f9
fix: Danenberg and Keren image io
