+++ tests/datasets/test_prepare_and_setup.py
@@ -32,10 +32,12 @@ def prepare_dataset(base_dir: Path, dataset_name: str):
     assert isinstance(ds.metadata, pd.DataFrame)
     assert isinstance(ds.intensity, pd.DataFrame)
 
-@pytest.mark.skip(reason="Only run when downloading changes. This will trigger the whole end-to-end runs.")
-@pytest.mark.parametrize("dataset_name", ["Jackson2020", "Danenberg2022", "Cords2024", "Keren2018"])
+# @pytest.mark.skip(reason="Only run when downloading changes. This will trigger the whole end-to-end runs.")
+# @pytest.mark.parametrize("dataset_name", ["Jackson2020", "Danenberg2022", "Cords2024", "Keren2018"])
+@pytest.mark.parametrize("dataset_name", ["Keren2018", "Jackson2020", "Danenberg2022", "Cords2024"])
 def test_prepare_data(dataset_name):
     tmpdir = Path(f'/Volumes/T7/ai4bmr-datasets/{dataset_name}')
+    tmpdir.mkdir(parents=True, exist_ok=True)
     if tmpdir is None:
         with tempfile.TemporaryDirectory() as tmpdir:
             base_dir = Path(tmpdir)
@@ -43,7 +45,7 @@ def test_prepare_data(dataset_name):
     else:
         prepare_dataset(base_dir=tmpdir, dataset_name=dataset_name)
 
+# test_prepare_data("Keren2018")
 # test_prepare_data("Jackson2020")
-# test_prepare_data("Danenberg2022")
-test_prepare_data("Keren2018")
+test_prepare_data("Danenberg2022")
 # test_prepare_data("Cords2024")
