### 2025-08-28 95f03ba
harmonize I/O
