+++ README.md
@@ -82,7 +82,7 @@ python3 -m venv .venv
 source .venv/bin/activate
 
 # Install the package
-pip install git+https://github.com/AI4SCR/ai4bmr-datasets.git@v2.0.4
+pip install git+https://github.com/AI4SCR/ai4bmr-datasets.git@v2.0.5
 ```
 
 #### Using `conda`
@@ -93,7 +93,7 @@ conda create -n ai4bmr-datasets python=3.12 -y
 conda activate ai4bmr-datasets
 
 # Install the package
-pip install git+https://github.com/AI4SCR/ai4bmr-datasets.git@v2.0.4
+pip install git+https://github.com/AI4SCR/ai4bmr-datasets.git@v2.0.5
 ```
 
 ---
