### 2025-08-28 e26d1e3
switch to poisx in R scripts
