### 2025-08-28 97f1327
fix: try new approach for window Rscript
