### 2025-08-29 aa8fb82
fix bugs and run all tests
