+++ tests/datasets/test_prepare_and_setup.py
@@ -1,9 +1,10 @@
-import pytest
 import tempfile
 from pathlib import Path
 
-def prepare_dataset(base_dir: Path, dataset_name: str):
+import pytest
 
+
+def prepare_dataset(base_dir: Path, dataset_name: str):
     match dataset_name:
         case "Jackson2020":
             from ai4bmr_datasets import Jackson2020 as Dataset
@@ -32,19 +33,14 @@ def prepare_dataset(base_dir: Path, dataset_name: str):
     assert isinstance(ds.metadata, pd.DataFrame)
     assert isinstance(ds.intensity, pd.DataFrame)
 
+
 @pytest.mark.skip(reason="Only run when downloading changes. This will trigger the whole end-to-end runs.")
 @pytest.mark.parametrize("dataset_name", ["Keren2018", "Jackson2020", "Danenberg2022", "Cords2024"])
 def test_prepare_data(dataset_name):
-    tmpdir = Path(f'/Volumes/T7/ai4bmr-datasets/{dataset_name}')
-    tmpdir.mkdir(parents=True, exist_ok=True)
+    tmpdir = None
     if tmpdir is None:
         with tempfile.TemporaryDirectory() as tmpdir:
             base_dir = Path(tmpdir)
             prepare_dataset(base_dir=base_dir, dataset_name=dataset_name)
     else:
         prepare_dataset(base_dir=tmpdir, dataset_name=dataset_name)
-
-# test_prepare_data("Keren2018")
-# test_prepare_data("Jackson2020")
-# test_prepare_data("Danenberg2022")
-# test_prepare_data("Cords2024")
