+++ src/ai4bmr_datasets/datasets/Danenberg.py
@@ -1,5 +1,4 @@
 import re
-import tempfile
 from pathlib import Path
 
 import numpy as np
@@ -391,6 +390,3 @@ class Danenberg2022(BaseIMCDataset):
         panel_path = self.get_panel_path("published")
         panel_path.parent.mkdir(parents=True, exist_ok=True)
         panel.to_parquet(panel_path)
-
-
-self = Danenberg2022()
