### 2025-08-29 dbc0bf1
update README and tidy repo
