### 2025-09-04 c1ad0ec
FIX: reset and sort panel index by mass
