### 2025-09-04 5674892
fix: reset panel index to avoid duplicated entries
