### 2025-09-29 404953b
updated trident and started building custom pipeline
