+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -29,22 +29,23 @@ from ai4bmr_learn.plotting.umap import plot_umap
 
 class BEAT:
     INVALID_SAMPLE_IDS = {
-        '1GUG.0M.2',  # barcode
-        '1GVQ.06.2',  # barcode
-        '1GUG.0M.0',  # mpp_x != mpp_y
-        '1GVQ.06.1',  # mpp_x != mpp_y
+        # '1GUG.0M.2',  # barcode
+        # '1GVQ.06.2',  # barcode
+        # '1GUG.0M.0',  # mpp_x != mpp_y
+        # '1GVQ.06.1',  # mpp_x != mpp_y
     }
+    PATH_TO_TRIDENT = Path("/users/mensmeng/workspace/dermatology/TRIDENT_pipeline/new_install/trident/trident")
 
     def __init__(self, base_dir: Path | None = None):
-        self.base_dir = base_dir or Path('/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat')
-        self.raw_dir = self.base_dir / '01_raw'
+        self.base_dir = base_dir or Path('/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat_rescanned')
+        self.raw_dir = Path('/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/mesothelioma/HnE/BEAT-MESO_rescan_40x_preXenium')
         self.processed_dir = self.base_dir / '02_processed'
         self.dataset_dir = self.processed_dir / 'datasets' / 'beat'
         self.tools_dir = self.base_dir / '03_tools'
 
         # GLOBAL PARAMS
         # self.target_mpp = 0.8624999831542972
-        self.target_mpp = 0.862515094
+        self.target_mpp = 0.
         self.overlap = 0.25
 
     def prepare_tools(self):
@@ -67,75 +68,26 @@ class BEAT:
         subprocess.run(install_cmd, shell=True, check=True)
 
     def prepare_metadata(self):
-        """Questions
-        Is this n=12 or n=14? I assume -14 can be ignored
-        2024-05-24_n.1 - n.14/malexan3-24-05-2024-012-14.czi
+        def filter_metadata(df: pd.DataFrame) -> pd.DataFrame:
+            df = df[df['tech_id'] == 'HE']
+            df = df[df['include'] == True]
 
-        What are the corresponding numbers for `2024-05-31_n98-n100`
+            ## remove columns where all values are NA
+            df = df.dropna(axis=1, how='all')
+            assert df.index.is_unique
+            assert df.sample_path.is_unique
+            assert df.sample_path.isna().sum() == 0
+            return df
+        
 
-        """
 
-        clinical = pd.read_excel(self.raw_dir / 'H_E_Images/Image_ID_vs_patient_blockIDs.xlsx')
-        clinical.columns = clinical.columns.map(tidy_name)
-        clinical = clinical.convert_dtypes()
-
-        assert clinical.sample_barcode.value_counts().max() == 1
-
-        dir_to_range = {
-            '2024-05-28 slide n.15 - n.50': list(range(15, 50 + 1)),
-            'Slide N°130-222': list(range(130, 222 + 1)),
-            '2024-05-24_n.1 - n.14': list(range(1, 14 + 1)),
-            '2024-05-30_n51 - n.97': list(range(51, 97 + 1)),
-            '2024-06-05_n101-n129': list(range(101, 129 + 1)),
-            '2024-05-31_n98-n100': list(range(98, 100 + 1))
-        }
-
-        import re
-        wsi_paths = list(self.base_dir.rglob('*.ndpi')) + list(self.base_dir.rglob('*.czi'))
-        records = []
-        for wsi_path in wsi_paths:
-
[…]
