+++ src/ai4bmr_datasets/utils/wsi/feats.py
@@ -1,0 +1,202 @@
+#%%
+import json
+import subprocess
+import textwrap
+from pathlib import Path
+
+# from dotenv import load_dotenv
+# load_dotenv('/users/mensmeng/workspace/dermatology/derm-pipeline/.env')
+
+
+import geopandas as gpd
+import numpy as np
+import openslide
+import pandas as pd
+#import tifffile
+import torch
+from tqdm import tqdm
+
+
+
+from openslide import OpenSlide
+import os
+
+#%%
+from trident import OpenSlideWSI
+from trident.segmentation_models.load import segmentation_model_factory
+from trident.IO import read_coords
+from trident.patch_encoder_models.load import encoder_factory
+from trident.slide_encoder_models.load import encoder_factory as slide_encoder_factory
+import h5py
+from dotenv import load_dotenv, dotenv_values
+from huggingface_hub import login
+from trident.slide_encoder_models.load import slide_to_patch_encoder_name
+from trident.IO import create_lock, remove_lock, is_locked, update_log
+
+
+#%%
+def check_fm_compatibility(target_mag, patch_size, feature_extractor, slide_extractor):
+
+    avail_feature_extractors = ['uni_v2', 'hoptimus0', 'virchow2', 'hibou_l', 'conch_v15', 'resnet50']
+    avail_slide_encoders = ['titan']
+
+
+    slide_to_patch_encoder_name = {
+    'threads': 'conch_v15',
+    'titan': 'conch_v15',
+    'tcga': 'conch_v15',
+    'prism': 'virchow',
+    'chief': 'ctranspath',
+    'gigapath': 'gigapath',
+    'madeleine': 'conch_v1',
+    'feather': 'conch_v15'
+    }
+
+    params_feature_extracor = {
+        'uni_v2': {'mag': 20, 'patch_size': 256},
+        'hoptimus0': {'mag': 20, 'patch_size': 224},
+        'virchow2': {'mag': 20, 'patch_size': 224},
+        'hibou_l': {'mag': 20, 'patch_size': 224},
+        'conch_v15': {'mag': 20, 'patch_size': 512},
+        'resnet50': {'mag': 20, 'patch_size': 256}
+    }
+
+    
+    if slide_extractor is not None:
+        if slide_extractor not in avail_slide_encoders:
+            raise ValueError(f"Slide encoder '{slide_extractor}' is not available. Please choose from {avail_slide_encoders}.")
+        
+    
+        if feature_extractor is None:
+            feature_extractor = slide_to_patch_encoder_name.get(slide_extractor)
+        else:
+            expected_extractor = slide_to_patch_encoder_name.get(slide_extractor)
+            if feature_extractor != expected_extractor:
+                raise ValueError(f"Feature extractor '{feature_extractor}' is not compatible with slide encoder '{slide_extractor}'. Please choose '{expected_extractor}' instead.")
+
+    if feature_extractor is not None:
+        if feature_extractor not in avail_feature_extractors:
+            raise ValueError(f"Feature extractor '{feature_extractor}' is not available. Please choose from {avail_feature_extractors}.")
+
+        expected_mag = params_feature_extracor.get(feature_extractor, {}).get('mag')
+        expected_patch_size = params_feature_extracor.get(feature_extractor, {}).get('patch_size')
+
+        if expected_mag is not None and target_mag != expected_mag:
+            raise ValueError(f"Target magnification '{target_mag}' is not compatible with feature extractor '{feature_extractor}'. Please choose '{expected_mag}' instead.")
+
+        if expected_patch_size is not None and patch_size != expected_patch_size:
[…]
