+++ pyproject.toml
@@ -9,7 +9,7 @@ authors = [
     { name = "Adriano Martinelli", email = "adriano.martinelli@chuv.ch" },
 ]
 readme = "README.md"
-requires-python = ">=3.12"
+requires-python = ">=3.10"
 dependencies = [
     "pydantic",
     "pydantic-settings",
