### 2025-10-08 e7656e7
patch feature analysis WIP
