+++ src/ai4bmr_datasets/utils/qc/patch_qc.py
@@ -1,0 +1,254 @@
+## file that contains the functions to extract features from patches (such as tissue proportion, intensity features, etc.)
+## can be saved as patch information in the anndata (with prefix 'patch_')
+
+import cv2
+import numpy as np
+import sys
+#from skimage.feature import graycomatrix, graycoprops
+from trident import OpenSlideWSI
+from pathlib import Path
+from sklearn.preprocessing import MinMaxScaler
+import pandas as pd
+import matplotlib.pyplot as plt
+from tqdm import tqdm  
+
+# %%
+def calculate_tissue_proportion(patch, visualize=False):
+    """
+    Calculate the proportion of tissue in a patch.
+    Args:
+        patch (np.ndarray): Image patch to analyze.
+        visualize (bool): Whether to visualize the binary mask.
+    Returns:
+        float: Proportion of tissue in the patch.
+    """
+    # Convert patch to grayscale
+    gray = cv2.cvtColor(patch, cv2.COLOR_BGR2GRAY)
+    
+    # Threshold the image to binary
+    _, binary = cv2.threshold(gray, 0, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
+    
+    # Calculate tissue proportion
+    tissue_proportion = np.sum(binary == 0) / (patch.shape[0] * patch.shape[1])
+    
+    if visualize:
+        plt.imshow(binary, cmap='gray')
+        plt.axis('off')
+        plt.title(f"Tissue proportion: {tissue_proportion:.2f}")
+        plt.show()
+    
+    return tissue_proportion
+# %%
+def calculate_basic_intensity_features(patch, visualize=False):
+    """
+    Calculate basic intensity features of a patch.
+    Args:
+        patch (np.ndarray): Image patch to analyze.
+        visualize (bool): Whether to visualize the patch.
+    Returns:
+        tuple: Mean, standard deviation, and median intensity values for each channel.
+    """
+    mean_intensity = np.mean(patch, axis=(0,1))  # Mean per channel
+    std_intensity = np.std(patch, axis=(0,1))
+    median_intensity = np.median(patch, axis=(0,1))
+
+    return mean_intensity, std_intensity, median_intensity
+
+## Needs to be explained and checked -> not yet ready for use
+# %%
+# def calculate_glcm_features(patch):
+#     # Convert patch to grayscale
+#     gray = cv2.cvtColor(patch, cv2.COLOR_BGR2GRAY)
+    
+#     # Calculate GLCM matrix
+#     glcm = graycomatrix(gray, [1], [0], 256, symmetric=True, normed=True)
+    
+#     # Calculate GLCM properties
+#     contrast = graycoprops(glcm, prop='contrast')[0, 0]
+#     dissimilarity = graycoprops(glcm, prop='dissimilarity')[0, 0]
+#     homogeneity = graycoprops(glcm, prop='homogeneity')[0, 0]
+#     energy = graycoprops(glcm, prop='energy')[0, 0]
+#     correlation = graycoprops(glcm, prop='correlation')[0, 0]
+    
+#     return contrast, dissimilarity, homogeneity, energy, correlation
+
+# %%
+def calculate_edge_density(patch, visualize=False):
+    """
+    Calculate the edge density of a patch using Canny edge detection.
+    Args:
+        patch (np.ndarray): Image patch to analyze.
+        visualize (bool): Whether to visualize the edges.
+    Returns:
+        float: Edge density of the patch.
+    """
+    # Convert patch to grayscale
+    gray = cv2.cvtColor(patch, cv2.COLOR_BGR2GRAY)
+    
+    # Calculate edges using Canny edge detector
[…]
