### 2025-10-14 57cd0cc
update beat embeddings
