+++ slurm/scripts/compute-beat-embeddings.sh
@@ -9,11 +9,11 @@
 #SBATCH --job-name=beat_embeddings_conch
 #SBATCH --output=/users/mensmeng/logs/%x-%A-%a.out
 #SBATCH --error=/users/mensmeng/logs/%x-%A-%a.err
-#SBATCH --array=0-23
+#SBATCH --array=2-178
 
 
-#SAMPLE_LIST="/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat_rescanned/02_processed/sample_ids.txt"
-SAMPLE_LIST="/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat_rescanned/02_processed/missing_files_conch.txt"
+SAMPLE_LIST="/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat_rescanned/02_processed/sample_ids.txt"
+#SAMPLE_LIST="/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat_rescanned/02_processed/missing_files_conch.txt"
 SAMPLE_ID=$(sed -n "$((SLURM_ARRAY_TASK_ID + 1))p" $SAMPLE_LIST)
 echo "Processing sample ID: $SAMPLE_ID"
 
