### 2025-10-21 eb6b30d
fix PCa metadata
