+++ src/ai4bmr_datasets/datasets/PCa.py
@@ -10,15 +10,51 @@ class PCa(BaseIMCDataset):
     id = "PCa"
     doi = "unpublished"
 
-    def __init__(self, base_dir: Path | None = None):
-        super().__init__(base_dir)
+    def __init__(self,
+                 base_dir: Path | None = None,
+                 image_version: str | None = None,
+                 mask_version: str | None = None,
+                 feature_version: str | None = None,
+                 metadata_version: str | None = None,
+                 load_intensity: bool = False,
+                 load_spatial: bool = False,
+                 load_metadata: bool = False,
+                 align: bool = False,
+                 join: str = "outer",
+                 ):
+        """
+        Initializes the Cords2024 dataset.
+
+        Args:
+            base_dir (Path | None): Base directory for the dataset.
+            image_version (str | None): Version of the images to load.
+            mask_version (str | None): Version of the masks to load.
+            feature_version (str | None): Version of the features to load.
+            metadata_version (str | None): Version of the metadata to load.
+            load_intensity (bool): Whether to load intensity features.
+            load_spatial (bool): Whether to load spatial features.
+            load_metadata (bool): Whether to load metadata.
+            align (bool): Whether to align sample IDs across modalities.
+            join (str): Type of join for alignment (e.g., 'outer').
+        """
+
+        super().__init__(base_dir=base_dir,
+                         image_version=image_version,
+                         mask_version=mask_version,
+                         feature_version=feature_version,
+                         metadata_version=metadata_version,
+                         load_intensity=load_intensity,
+                         load_spatial=load_spatial,
+                         load_metadata=load_metadata,
+                         align=align,
+                         join=join)
 
         # raw paths
         self.raw_clinical_metadata_path = (
                 self.base_dir / "01_raw" / "metadata" / "ROI_matching_blockID.xlsx"
         )
         self.raw_tma_annotations_path = (
-                self.base_dir / "01_raw" / "metadata" / "tma-annotations-v2.xlsx"
+                self.base_dir / "01_raw" / "metadata" / "tma-annotations-v3.xlsx"
         )
 
         self.raw_tma_tidy_path = self.base_dir / "01_raw" / "metadata" / "TMA_tidy.txt"
@@ -491,7 +527,7 @@ class PCa(BaseIMCDataset):
         mapping = mapping[~filter_]
 
         sample_ids = set([i.stem for i in (self.raw_dir / 'acquisitions').glob('*.tiff')])
-        file_name_to_sample_id = {re.sub(r'\d+$', '', i) + ".tiff": i for i in sample_ids}
+        file_name_to_sample_id = {re.sub(r'_\d+$', '', i) + ".tiff": i for i in sample_ids}
         assert len(file_name_to_sample_id) == len(sample_ids)
         assert set(mapping.file_name_tidy) == set(file_name_to_sample_id.keys())
 
@@ -613,7 +649,7 @@ class PCa(BaseIMCDataset):
         )
 
         mapping = self.get_napari_file_name_to_sample_id_mapping()
-        sample_ids = roi_match_blockId.file_name_napari.map(mapping).to_list()
+        sample_ids = roi_match_blockId.file_name_napari.map(mapping)
 
         roi_match_blockId = roi_match_blockId.assign(
             sample_id=sample_ids,
@@ -622,14 +658,12 @@ class PCa(BaseIMCDataset):
[…]
