### 2025-10-29 bc0e6b8
Add community dataset contribution guidelines
