+++ .github/pull_request_template.md
@@ -1,0 +1,24 @@
+## Summary
+- [ ] This PR adds or updates documentation only.
+- [ ] This PR introduces or modifies dataset processing code.
+- [ ] This PR changes core library behaviour.
+- [ ] Other (describe below).
+
+> Provide a concise overview of the changes and reference related issues (e.g., `Fixes #123`).
+
+## Dataset contributions (if applicable)
+- Dataset name and DOI/source:
+- Hosted at (Figshare/Zenodo/other):
+- [ ] Raw multiplexed images are publicly accessible.
+- [ ] Segmentation masks are included or justified if omitted.
+- [ ] Published intensity/feature tables are provided when available.
+- [ ] Module inherits from `BaseIMCDataset` and registers the dataset in the package `__init__` files.
+- [ ] README and documentation entries for the dataset were updated.
+
+## Testing
+List the commands you executed to validate the change, for example:
+- `pytest tests/datasets -k NewDataset`
+- `pytest`
+- `ruff check`
+
+If automated tests are not available, explain the manual validation that was performed.
