### 2025-10-30 d81b3da
Refactor download records to include metadata
