### 2025-11-19 #8780 @ujjwalkarn
@adrianomartinelli @rapsoman could you please take a look at the two open issues (from @Timozen's comment above) so we can move the review forward, thank you!

@tensorsofthewall wondering if you have any update on the review? Thanks!
