+++ tests/datasets/test_Keren2018.py
@@ -8,7 +8,7 @@ def test_Keren2018():
     _dir = Path('/Volumes/T7')
     _dir = _dir if _dir.exists() else None
     with tempfile.TemporaryDirectory(dir=_dir) as tmpdir:
-        base_dir = Path(tmpdir) / 'Keren2018'
+        base_dir = Path(tmpdir) / 'ai4bmr-datasets' / 'Keren2018'
         ds = self = Keren2018(base_dir=base_dir, image_version=image_version, mask_version=mask_version,
                               metadata_version=metadata_version, load_metadata=True,
                               feature_version=feature_version, load_intensity=True, load_spatial=True)
