### 2025-11-24 7fa146a
helper to compute all checksums
