### 2025-11-25 161815d
collect checksums
