+++ scripts/compute-checksums-for-all-datasets.py
@@ -1,15 +1,15 @@
 from dotenv import load_dotenv
 assert load_dotenv()
-
 import hashlib
 from pathlib import Path
 import os
 import pandas as pd
+from tqdm import tqdm
 
 dataset_dirs = [i for i in Path(os.getenv("AI4BMR_DATASETS_DIR")).iterdir() if i.is_dir()]
 
 records = []
-for dir_ in dataset_dirs:
+for dir_ in tqdm(dataset_dirs):
     zip_files = list(dir_.rglob("*.zip"))
     for zip_file in zip_files:
         # Compute SHA‑256 directly from the ZIP file
@@ -27,3 +27,5 @@ for dir_ in dataset_dirs:
 
 df = pd.DataFrame.from_records(records)
 print(df.to_markdown())
+save_path = Path().resolve() / 'checksums.md'
+df.to_markdown(save_path, index=False)
