### 2025-11-25 0fcfbd4
Add checksums to dataset download records
