+++ src/ai4bmr_datasets/datasets/Cords2024.py
@@ -107,90 +107,112 @@ class Cords2024(BaseIMCDataset):
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/2020115_LC_NSCLC_TMA_86_A.mcd.zip?download=1",
                 file_name="2020115_LC_NSCLC_TMA_86_A.mcd.zip",
+                checksum="7bf62d8252ad15bcf11a34dc64294f72e6618ac30ef8b17f5c39b6bd4760322f",
             ),
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/2020117_LC_NSCLC_TMA_86_B.mcd.zip?download=1",
                 file_name="2020117_LC_NSCLC_TMA_86_B.mcd.zip",
+                checksum="dbec22705ed3f4811222aa253d173eb167111d85b05bf018608c57557a114a1a",
             ),
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/2020120_LC_NSCLC_TMA_86_C.mcd.zip?download=1",
                 file_name="2020120_LC_NSCLC_TMA_86_C.mcd.zip",
+                checksum="850db4fd90d0b101667b2595c757bf0a8744bb4234cb637e4d9b84f6d661f6c5",
             ),
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/20201219_LC_NSCLC_TMA_178_A.mcd.zip?download=1",
                 file_name="20201219_LC_NSCLC_TMA_178_A.mcd.zip",
+                checksum="ff97fa311ef948da9f245abdb9bd771616e1996563d5f4a417f2b8d5d8a2de02",
             ),
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/20201222_LC_NSCLC_TMA_178_B_2.mcd.zip?download=1",
                 file_name="20201222_LC_NSCLC_TMA_178_B_2.mcd.zip",
+                checksum="3e41a7e21fb63581609bc4e6962076849a876336405cbd25336213b3ecec1da8",
             ),
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/20201224_LC_NSCLC_TMA_178_C.mcd.zip?download=1",
                 file_name="20201224_LC_NSCLC_TMA_178_C.mcd.zip",
+                checksum="1a2b2ec79488340a0a54a746d1206b61a94bc15182ebcd5fe3b4243fd3ddb523",
             ),
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/20201226_LC_NSCLC_TMA_175_A.mcd.zip?download=1",
                 file_name="20201226_LC_NSCLC_TMA_175_A.mcd.zip",
+                checksum="86f47ecce84c07b2162e682ffe7effd1af703918c0d2fbe5c738a4817e9c3c70",
             ),
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/20201228_LC_NSCLC_TMA_175_B.mcd.zip?download=1",
                 file_name="20201228_LC_NSCLC_TMA_175_B.mcd.zip",
+                checksum="0f02254e85adcac25a1aeb91ac390590dd757037e18799ea4ff664369d023e72",
             ),
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/20210101_LC_NSCLC_TMA_175_C.mcd.zip?download=1",
                 file_name="20210101_LC_NSCLC_TMA_175_C.mcd.zip",
+                checksum="13885eda04dee6d5718a13e940b0af5699af5263812733d14458c7bac7025fde",
             ),
             DownloadRecord(
                 url="https://zenodo.org/records/7961844/files/20210104_LC_NSCLC_TMA_176_A.mcd.zip?download=1",
[…]
