+++ tests/datasets/test_prepare_and_setup.py
@@ -1,3 +1,8 @@
+import os
+
+from dotenv import load_dotenv
+load_dotenv()
+
 import tempfile
 from pathlib import Path
 
@@ -44,3 +49,23 @@ def test_prepare_data(dataset_name):
             prepare_dataset(base_dir=base_dir, dataset_name=dataset_name)
     else:
         prepare_dataset(base_dir=tmpdir, dataset_name=dataset_name)
+
+
+@pytest.mark.parametrize("dataset_name", ["Keren2018", "Jackson2020", "Danenberg2022", "Cords2024"])
+def test_checksums(dataset_name):
+    match dataset_name:
+        case "Jackson2020":
+            from ai4bmr_datasets import Jackson2020 as Dataset
+        case "Danenberg2022":
+            from ai4bmr_datasets import Danenberg2022 as Dataset
+        case "Cords2024":
+            from ai4bmr_datasets import Cords2024 as Dataset
+        case "Keren2018":
+            from ai4bmr_datasets import Keren2018 as Dataset
+        case _:
+            raise ValueError(f"Unknown dataset name: {dataset_name}")
+
+    base_dir = os.getenv('')
+    ds = Dataset(base_dir=base_dir,
+                 image_version='',
+                 mask_version='')
