### 2025-11-25 d38eea3
run checksum tests
