### 2025-11-25 5dc8389
unlink if checksum fails
