### 2025-11-25 fd24a2c
use pyteet tmp dir
