+++ src/ai4bmr_datasets/datasets/BEAT.py
@@ -1,651 +1,0 @@
-import json
-import subprocess
-import textwrap
-from pathlib import Path
-
-import geopandas as gpd
-import numpy as np
-import openslide
-import pandas as pd
-import tifffile
-import torch
-from ai4bmr_core.utils.tidy import tidy_name
-from ai4bmr_learn.data_models.Coordinate import SlideCoordinate
-from ai4bmr_learn.datasets.items import SlidePatches, get_patch
-from ai4bmr_learn.plotting.patches import visualize_coords
-from ai4bmr_learn.utils.device import batch_to_device
-from ai4bmr_learn.utils.images import filter_coords
-from ai4bmr_learn.utils.slides import get_coordinates_dict, get_mpp, get_mpp_and_resolution, get_seg_model, \
-    get_slide_patcher_params, get_thumbnail, segment_slide
-from loguru import logger
-from openslide import OpenSlide
-from skimage.io import imsave
-from torch.utils.data import DataLoader
-from torchvision.transforms import InterpolationMode, v2
-from tqdm import tqdm
-from trident.patch_encoder_models import encoder_factory
-from ai4bmr_learn.plotting.umap import plot_umap
-
-
-class BEAT:
-    INVALID_SAMPLE_IDS = {
-        '1GUG.0M.2',  # barcode
-        '1GVQ.06.2',  # barcode
-        '1GUG.0M.0',  # mpp_x != mpp_y
-        '1GVQ.06.1',  # mpp_x != mpp_y
-    }
-
-    def __init__(self, base_dir: Path | None = None):
-        self.base_dir = base_dir or Path('/work/PRTNR/CHUV/DIR/rgottar1/spatial/data/beat')
-        self.raw_dir = self.base_dir / '01_raw'
-        self.processed_dir = self.base_dir / '02_processed'
-        self.dataset_dir = self.processed_dir / 'datasets' / 'beat'
-        self.tools_dir = self.base_dir / '03_tools'
-
-        # GLOBAL PARAMS
-        # self.target_mpp = 0.8624999831542972
-        self.target_mpp = 0.862515094
-        self.overlap = 0.25
-
-    def prepare_tools(self):
-        bfconvert = self.tools_dir / "bftools" / "bfconvert"
-
-        if bfconvert.exists():
-            logger.info(f"bfconvert found at {bfconvert}. Skipping installation.")
-            return
-
-        logger.info("⚙️Installing Bio-Formats tools.")
-        self.tools_dir.mkdir(parents=True, exist_ok=True)
-
-        install_cmd = f"""
-        wget https://downloads.openmicroscopy.org/bio-formats/8.2.0/artifacts/bftools.zip -O {self.tools_dir}/bftools.zip
-        unzip {self.tools_dir}/bftools.zip -d {self.tools_dir}
-        {self.tools_dir / "bftools" / "bfconvert"} -version || exit
-        """
-
-        install_cmd = textwrap.dedent(install_cmd).strip()
-        subprocess.run(install_cmd, shell=True, check=True)
-
-    def prepare_metadata(self):
-        """Questions
-        Is this n=12 or n=14? I assume -14 can be ignored
-        2024-05-24_n.1 - n.14/malexan3-24-05-2024-012-14.czi
-
-        What are the corresponding numbers for `2024-05-31_n98-n100`
-
-        """
-
-        clinical = pd.read_excel(self.raw_dir / 'H_E_Images/Image_ID_vs_patient_blockIDs.xlsx')
-        clinical.columns = clinical.columns.map(tidy_name)
-        clinical = clinical.convert_dtypes()
-
-        assert clinical.sample_barcode.value_counts().max() == 1
-
[…]
