+++ src/ai4bmr_datasets/datasets/__init__.py
@@ -4,10 +4,3 @@ from .Danenberg import Danenberg2022
 from .Jackson2020 import Jackson2020
 
 from .PCa import PCa
-from .BEAT import BEAT
-
-from .BEAT import BEAT
-
-# from .DummyPoints import DummyPoints
-# from .DummyImages import DummyImages
-# from .DummyTabular import DummyTabular
