### 2026-01-21 #8780 @tensorsofthewall
Nearly done! I'm currently struggling with some machine-side issues related to installation and functionality (not from the library side, my machine is a little old and finnicky). But I'll have that sorted. Finished most items on my checklist, software paper and documentation looks good. Will update in 2-3 days with any issues directly on the repo, otherwise will have my review complete by Saturday
