### 2026-01-29 AI4SCR/ai4bmr-datasets#33 @tensorsofthewall
JOSS - Errors when running example code
Hi 👋
As part of the ongoing **JOSS review**, I attempted to run the example code provided in the documentation and ran into two issues. I’m documenting them here to make debugging and fixing easier.

---

## Environment

- **OS**:
  - Windows 11
  - WSL2 (Ubuntu)
- **Python**: 3.12
- **Environment manager**: `conda` / `micromamba`
- **Installation**: followed instructions in the README
- **Library version**: 2.1.1
---

## Issue 1: `fastparquet` fails when preparing the Keren2018 dataset

### Reproduction code

```python
from ai4bmr_datasets import Keren2018
from pathlib import Path

# Define the base directory for storing the dataset
base_dir = Path('/mnt/d/projects/ai4bmr/datasets')

# Initialize the dataset
dataset = Keren2018(
    base_dir=base_dir,
    image_version="published",
    mask_version="published",
    feature_version="published",
    load_intensity=True,
    metadata_version="published",
    load_metadata=True
)

# Download and preprocess the data
dataset.prepare_data()
```

### Error
```text
2026-01-29 08:05:10.074 | INFO     | ai4bmr_datasets.datasets.Keren2018:create_panel:95 - Creating panel
Traceback (most recent call last):
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/fastparquet/writer.py", line 290, in convert
    out = array_encode_utf8(data)
          ^^^^^^^^^^^^^^^^^^^^^^^
  File "fastparquet/speedups.pyx", line 42, in fastparquet.speedups.array_encode_utf8
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/pandas/core/series.py", line 901, in __array__
    arr = np.array(values, dtype=dtype, copy=copy)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/pandas/core/arrays/arrow/array.py", line 855, in __array__
    raise ValueError(
ValueError: Unable to avoid copy while creating an array as requested.

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/ai4bmr_datasets/datasets/Keren2018.py", line 35, in prepare_data
    self.create_panel()
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/ai4bmr_datasets/datasets/Keren2018.py", line 134, in create_panel
    panel.to_parquet(panel_path, engine='fastparquet')
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/pandas/core/frame.py", line 3135, in to_parquet
    return to_parquet(
           ^^^^^^^^^^^
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/pandas/io/parquet.py", line 490, in to_parquet
    impl.write(
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/pandas/io/parquet.py", line 342, in write
    self.api.write(
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/fastparquet/writer.py", line 1340, in write
    write_simple(filename, data, fmd,
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/fastparquet/writer.py", line 1001, in write_simple
    write_to_file(f)
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/fastparquet/writer.py", line 985, in write_to_file
    rg = make_row_group(f, row_group, fmd.schema,
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/fastparquet/writer.py", line 802, in make_row_group
    chunk = write_column(f, coldata, column,
            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/fastparquet/writer.py", line 635, in write_column
    repetition_data, definition_data, encode[encoding](data, selement), 8 * b'\x00'
                                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/fastparquet/writer.py", line 388, in encode_plain
    out = convert(data, se)
          ^^^^^^^^^^^^^^^^^
  File "/home/san2597/micromamba/envs/ai4bmr-datasets/lib/python3.12/site-packages/fastparquet/writer.py", line 298, in convert
    raise ValueError('Error converting column "%s" to bytes using '
ValueError: Error converting column "target" to bytes using encoding UTF8. Original error: Unable to avoid copy while creating an array as requested.
```

I'm able to resolve this error by adding `import pandas as pd; pd.options.mode.string_storage = "python"` before calling dataset.prepare().

I do get several warnings like
```text
2026-01-29 08:11:22.216 | WARNING  | ai4bmr_datasets.datasets.BaseIMCDataset:create_annotated:342 - 28 has 6061 common objects and 12 mask objects without metadata
```

## Issue 2: `dataset.sample_ids` is [None, None, ...] after setup
After preparing the dataset, I moved on to Step 2: Setup and Access Data.
```
# List available data versions
dataset.list_image_versions()
dataset.list_mask_versions()
dataset.list_metadata_versions()
dataset.list_features_versions()

# Setup the dataset with a specific version and load features
dataset.setup()

# Access core components of the dataset
print("Sample IDs:", dataset.sample_ids)
print("Available images:", list(dataset.images.keys()))
print("Available masks:", list(dataset.masks.keys()))
```
Output:
```
Sample IDs: [None, None, None, None, None, None, None, None, None, None, None,
             None, None, None, None, None, None, None, None, None, None, None,
             None, None, None, None, None, None, None, None, None, None, None,
             None, None, None, None, None, None, None, None]

Available images: ['1', '10', '11', ..., '9']
Available masks:  ['1', '10', '11', ..., '9']
```

Since the sample IDs are all none, I can't access image/mask data in the next step(s).
