### 2026-01-29 #8780 @tensorsofthewall
Hi @adrianomartinelli ,
I just raised an issue (https://github.com/AI4SCR/ai4bmr-datasets/issues/33), having some trouble with the example code provided in the readme. Once that's resolved, we can proceed.
