### 2026-01-29 466db8c
pin pandas and pyarrow
