+++ pyproject.toml
@@ -14,15 +14,14 @@ dependencies = [
     "pydantic",
     "pydantic-settings",
     "tqdm",
-    "numpy",
+    "numpy<2.4",
     "openpyxl",
-    "pandas",
+    "pandas<3",
     "tifffile",
     "imagecodecs",
-    "zarr>3",
     "scikit-image",
     "matplotlib",
-    "pyarrow",
+    "pyarrow<22.0.0",
     "fastparquet",
     "geopandas",
     "loguru",
