### 2026-01-29 #8780 @adrianomartinelli
Hei @tensorsofthewall ,

thanks! just had a look and I think its simply a version mismatch between arrow/pandas. I pinned the versions in the latest version on main. Could you try to reinstall the package in a fresh env as described in issue?

Thanks,
Adriano
