### 2026-01-30 #8780 @adrianomartinelli
Hei @tensorsofthewall, new release is published 👍🏼.

Best,
Adriano
