### 2026-02-24 #8780 @adrianomartinelli
Hei @ujjwalkarn,

wanted to quickly check with you whether you need anything from us to move forward with the review?

Best,
Adriano
