### 2026-02-25 #8780 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
