### 2026-03-06 #8780 @crvernon
@adrianomartinelli - I just need you to address the following before I accept:

In your paper:

- [ ] Please remove the bold formatting from the title of your paper.
- [ ] LINE 36:  Missing space between "Zenodo" and "(European Organization For Nuclear..."

That's it.  Just let me know when these changes have been made and I'll accept.
